### 2021-05-04 #3239 @whedon
[PRE REVIEW]: radioactivedecay: A Python package for radioactive decay calculations
**Submitting author:** @alexmalins (<a href="http://orcid.org/0000-0003-1922-4496">Alex Malins</a>)
**Repository:** <a href="https://github.com/alexmalins/radioactivedecay/" target ="_blank">https://github.com/alexmalins/radioactivedecay/</a>
**Version:** v0.3.2
