### 2021-05-14 cd0d7b4
For release 0.3.3
Add MANIFEST.in, update LICENSE year
