### 2021-05-17 c13b49c
Document method for installing via conda
