### 2021-05-28 #3318 @whedon
[REVIEW]: radioactivedecay: A Python package for radioactive decay calculations
**Submitting author:** @alexmalins (<a href="http://orcid.org/0000-0003-1922-4496">Alex Malins</a>)
**Repository:** https://github.com/alexmalins/radioactivedecay/
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.4.9
