### 2021-06-10 #3318 @kellyrowland
Hi @shyamd 👋 - would you be available and willing to review this paper for JOSS?
