### 2021-06-11 #3318 @shyamd
> Hi @shyamd 👋 - would you be available and willing to review this paper for JOSS?

Sounds good!
