### 2021-06-11 alexmalins/radioactivedecay#31 @shyamd
JOSS Paper: State of Field
Hi Alex,

Can you add a few sentences on what challenges a user might face using `pyNE` or `Radiological toolbox` that `radioactivedecay` aims to solve. It's sort of hinted at in the capabilities and validation, but it would be useful to be very clear about this, maybe at the end of the `Statement of Need` section.
