### 2021-06-14 1ba6ca4
Update JoSS paper Statement of Need section
