+++ paper/paper.bib
@@ -9,6 +9,17 @@ year = {2010},
 doi = {10.1088/0952-4746/26/3/N02}
 }
 
+@article{Bakin2018,
+author = {Bakin, R. I. and Kiselev, A. A. and Shvedov, A. M. and Shikin, A. V.},
+title = {{Computational Errors in the Calculation of Long Radioactive Decay Chains}},
+journal = {Atomic Energy},
+volume = {123},
+number = {6},
+pages = {406--411},
+year = {2018},
+doi = {10.1007/s10512-018-0360-2}
+}
+
 @techreport{Endo2005,
 author = {Endo, Akira and Yamaguchi, Yasuhiro and Eckerman, Keith F.},
 title = {{JAERI 1347: Nuclear Decay Data for Dosimetry Calculation; Revised data of ICRP Publication 38}},
