### 2021-06-14 alexmalins/radioactivedecay#31 @alexmalins
Hi Shyam,

Thanks for this comment and for starting your review. I added a new para into the `Statement of Need` section accordingly (https://github.com/alexmalins/radioactivedecay/commit/1ba6ca40985e88271cd8c8af1046b8cf3a0e2fb0). Happy hear your thoughts and any further comments if you have them.
