### 2021-06-14 alexmalins/radioactivedecay#31 @shyamd
Thanks! That is perfect.
