### 2021-06-14 #3318 @shyamd
@alexmalins took care of my only concern promptly in https://github.com/alexmalins/radioactivedecay/issues/31 and I think my review is finished.
