### 2021-06-14 alexmalins/radioactivedecay#31 @alexmalins
Thanks @shyamd
