### 2021-07-09 2c4ee43
Added new input types, contents now holds numbers
