+++ radioactivedecay/inventory.py
@@ -1,6 +1,6 @@
 """
 The inventory module defines the ``Inventory`` class. Each ``Inventory`` instance contains one or
-more radionuclides, each with an associated activity. The decay of the radionuclide(s) in an
+more radionuclides, each with an associated number. The decay of the radionuclide(s) in an
 ``Inventory`` can be calculated by using the ``decay()`` method (normal double-precision
 floating-point operations). Use the ``decay_high_precision()`` method to perform a SymPy high
 numerical precision calculation. A ``DecayData`` dataset associated with each ``Inventory``
@@ -20,6 +20,7 @@ from functools import singledispatch, update_wrapper
 from typing import Callable, Dict, List, Tuple, Union
 import matplotlib
 from sympy import exp, nsimplify
+from scipy.constants import Avogadro
 from radioactivedecay.decaydata import DecayData, DEFAULTDATA, np
 from radioactivedecay.plots import _decay_graph
 from radioactivedecay.radionuclide import Radionuclide
@@ -37,20 +38,20 @@ def _add_dictionaries(
     dict_a: Dict[str, float], dict_b: Dict[str, float]
 ) -> Dict[str, float]:
     """
-    Adds together two dictionaries of radionuclides and associated activities.
+    Adds together two dictionaries of radionuclides and associated numbers.
 
     Parameters
     ----------
     dict_a : dict
-        First dictionary containing radionuclide strings as keys and activities as values.
+        First dictionary containing radionuclide strings as keys and numbers as values.
     dict_b : dict
-        Second dictionary containing radionuclide strings as keys and activities as values.
+        Second dictionary containing radionuclide strings as keys and numbers as values.
 
     Returns
     -------
     dict
         Combined dictionary containing the radionuclides in both dict_a and dict_b, where
-        activities have been added together when a radionuclide is present in both input
+        numbers have been added together when a radionuclide is present in both input
         dictionaries.
 
     Examples
@@ -63,13 +64,225 @@ def _add_dictionaries(
     """
 
     new_dict = dict_a.copy()
-    for radionuclide, radioactivity in dict_b.items():
+    for radionuclide, number in dict_b.items():
         if radionuclide in new_dict:
-            new_dict[radionuclide] = new_dict[radionuclide] + radioactivity
+            new_dict[radionuclide] = new_dict[radionuclide] + number
         else:
-            new_dict[radionuclide] = radioactivity
+            new_dict[radionuclide] = number
 
     return new_dict
+
+
+def _activity_to_number(
+    radionuclide: str,
+    activity : float,
+    data: DecayData = DEFAULTDATA
+) -> float:
+    """
+    Converts activity to number of atoms.
+
+    Parameters
+    ----------
+    radionuclide : str
+        The name of the material to be converted.
+    activity : float
+        The activity of the material to be converted.
+    data : DecayData, optional
+        Decay dataset (default is the ICRP-107 dataset).
+
+    Returns
+    -------
+    float
+        Number of atoms of material.
+
+    Example
+    --------
+    >>> rd.inventory._activity_to_number("C-14", 250)
+    64876004584874.73
+
+    """
+
+    try:
+        index = data.radionuclide_dict[radionuclide]
+    except:
+        raise ValueError(
+            radionuclide
+            + " is not a valid member of "
+            + data.dataset
+            )
+        
+    return (activity / data.scipy_data.decay_consts[index])
+
+
+def _mass_to_number(
+    radionuclide: str,
[…]
