### 2021-07-13 8b0d729
Updated tests to pass with stable nuclide data
