+++ tests/test_decaydata.py
@@ -187,7 +187,7 @@ class Test(unittest.TestCase):
             "DecayMatrices: data stored in SymPy objects for arbitrary-precision calculations.",
         )
 
-    def test_decaydata_instantiation(self):
+    def test_decaydata_instantiation(self):#hello
         """
         Test instantiation of DecayData objects.
         """
@@ -199,22 +199,21 @@ class Test(unittest.TestCase):
         self.assertEqual(data.dataset, "icrp107")
         self.assertEqual(data.hldata[0][0], 100.5)
         self.assertEqual(data.hldata[0][1], "d")
-        self.assertEqual(data.hldata[-1][0], 12.32)
-        self.assertEqual(data.hldata[-1][1], "y")
-        self.assertEqual(data.num_radionuclides, 1252)
+        self.assertEqual(data.hldata[-1][0], np.inf)
+        self.assertEqual(data.hldata[-1][1], "s")
+        self.assertEqual(data.num_radionuclides, 1498)
         self.assertEqual(data.radionuclides[0], "Fm-257")
-        self.assertEqual(data.radionuclides[-1], "H-3")
+        self.assertEqual(data.radionuclides[-1], "He-3")
         self.assertEqual(data.radionuclide_dict["Fm-257"], 0)
-        self.assertEqual(data.radionuclide_dict["H-3"], 1251)
+        self.assertEqual(data.radionuclide_dict["He-3"], 1497)
         self.assertEqual(list(data.prog_bfs_modes[0].keys())[0], "Cf-253")
         self.assertEqual(list(data.prog_bfs_modes[0].values())[0][0], 0.9979)
         self.assertEqual(list(data.prog_bfs_modes[0].values())[0][1], "\u03b1")
         self.assertEqual(list(data.prog_bfs_modes[0].keys())[1], "SF")
         self.assertEqual(list(data.prog_bfs_modes[0].values())[1][0], 0.0021)
         self.assertEqual(list(data.prog_bfs_modes[0].values())[1][1], "SF")
-        self.assertEqual(list(data.prog_bfs_modes[-1].keys())[0], "He-3")
-        self.assertEqual(list(data.prog_bfs_modes[-1].values())[0][0], 1.0)
-        self.assertEqual(list(data.prog_bfs_modes[-1].values())[0][1], "\u03b2-")
+        self.assertEqual(list(data.prog_bfs_modes[-1].keys()), [])
+        self.assertEqual(list(data.prog_bfs_modes[-1].values()), [])
         self.assertEqual(
             data.scipy_data.decay_consts[0], np.log(2) / (100.5 * 24 * 60 * 60)
         )
@@ -231,22 +230,22 @@ class Test(unittest.TestCase):
         self.assertEqual(data.dataset, "icrp107_2")
         self.assertEqual(data.hldata[0][0], 100.5)
         self.assertEqual(data.hldata[0][1], "d")
-        self.assertEqual(data.hldata[-1][0], 12.32)
-        self.assertEqual(data.hldata[-1][1], "y")
-        self.assertEqual(data.num_radionuclides, 1252)
+        self.assertEqual(data.hldata[-1][0], np.inf)
+        self.assertEqual(data.hldata[-1][1], "s")
+        self.assertEqual(data.num_radionuclides, 1498)
         self.assertEqual(data.radionuclides[0], "Fm-257")
-        self.assertEqual(data.radionuclides[-1], "H-3")
+        self.assertEqual(data.radionuclides[-1], "He-3")
         self.assertEqual(data.radionuclide_dict["Fm-257"], 0)
[…]
