### 2021-07-16 6f83bfd
Fixed tests to pass for new API, inventory.py bugs
