+++ tests/test_inventory.py
@@ -53,50 +53,49 @@ class Test(unittest.TestCase):
         radionuclides = ["H-3", "C-14"]
         H3 = Radionuclide("H3")
         C14 = Radionuclide("C14")
-        dataset = "test"
 
         # Dictionary parsing
         self.assertEqual(
-            _check_dictionary({"H-3": 1.0}, radionuclides, dataset), {"H-3": 1.0}
+            _check_dictionary({"H-3": 1.0}, radionuclides, input_type="number"), {"H-3": 1.0}
         )
         self.assertEqual(
-            _check_dictionary({"H3": 1.0}, radionuclides, dataset), {"H-3": 1.0}
+            _check_dictionary({"H3": 1.0}, radionuclides, input_type="number"), {"H-3": 1.0}
         )
         self.assertEqual(
-            _check_dictionary({"3H": 1.0}, radionuclides, dataset), {"H-3": 1.0}
+            _check_dictionary({"3H": 1.0}, radionuclides, input_type="number"), {"H-3": 1.0}
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1}, radionuclides, dataset), {"H-3": 1}
+            _check_dictionary({"H-3": 1}, radionuclides, input_type="number"), {"H-3": 1}
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1}, radionuclides, dataset), {"H-3": 1.0}
+            _check_dictionary({"H-3": 1}, radionuclides, input_type="number"), {"H-3": 1.0}
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1.0, "C-14": 2.0}, radionuclides, dataset),
+            _check_dictionary({"H-3": 1.0, "C-14": 2.0}, radionuclides, input_type="number"),
             {"H-3": 1.0, "C-14": 2.0},
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1.0, "C-14": 2.0}, radionuclides, dataset),
+            _check_dictionary({"H-3": 1.0, "C-14": 2.0}, radionuclides, input_type="number"),
             {"C-14": 2.0, "H-3": 1.0},
         )
         self.assertEqual(
-            _check_dictionary({H3: 1.0, C14: 2.0}, radionuclides, dataset),
+            _check_dictionary({H3: 1.0, C14: 2.0}, radionuclides, input_type="number"),
             {"C-14": 2.0, "H-3": 1.0},
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1.0, C14: 2.0}, radionuclides, dataset),
+            _check_dictionary({"H-3": 1.0, C14: 2.0}, radionuclides, input_type="number"),
             {"C-14": 2.0, "H-3": 1.0},
         )
         self.assertEqual(
-            _check_dictionary({H3: 1.0, "C-14": 2.0}, radionuclides, dataset),
+            _check_dictionary({H3: 1.0, "C-14": 2.0}, radionuclides, input_type="number"),
             {"C-14": 2.0, "H-3": 1.0},
         )
 
         # Catch incorrect arguments
         with self.assertRaises(ValueError):
-            _check_dictionary({"H-3": "1.0"}, radionuclides, dataset)
+            _check_dictionary({"H-3": "1.0"}, radionuclides, input_type="number")
         with self.assertRaises(ValueError):
-            _check_dictionary({"1.0": "H-3"}, radionuclides, dataset)
[…]
