### 2021-07-19 fccc405
Updated input types to be plural for consistency
