+++ radioactivedecay/inventory.py
@@ -352,12 +352,12 @@ def _input_to_number(
     
     number_abundance_dict, number_dict = {}, {}
     
-    if input_type == "mass":
+    if input_type == "masses":
         for nuc, mass in input_inv_dict.items():
             number = _mass_to_number(nuc, mass, data)
             number_dict[nuc] = number
         converted_dict = number_dict.copy()
-    elif input_type == "activity":
+    elif input_type == "activities":
         for nuc, act in input_inv_dict.items():
             number = _activity_to_number(nuc, act, data)
             number_dict[nuc] = number
@@ -379,7 +379,7 @@ def _input_to_number(
 def _check_dictionary(
     input_inv_dict: Dict[Union[str, Radionuclide], float],
     radionuclides: List[str],
-    input_type: str = "activity",
+    input_type: str = "activities",
     data: DecayData = DEFAULTDATA
 ) -> Dict[str, float]:
     """
@@ -394,7 +394,7 @@ def _check_dictionary(
     radionuclides : List[str]
         List of all the radionuclides in the decay dataset.
     input_type : str, optional
-        Type of input (number, mass (abundance), or activity), default is activity.
+        Type of input (numbers, masses (abundance), or activities), default is activities.
     data : DecayData, optional
         Decay dataset (default is the ICRP-107 dataset).
 
@@ -411,10 +411,10 @@ def _check_dictionary(
 
     Examples
     --------
-    >>> rd.inventory._check_dictionary({'3H': 1.0}, rd.DEFAULTDATA.radionuclides, "number", rd.DEFAULTDATA.dataset)
+    >>> rd.inventory._check_dictionary({'3H': 1.0}, rd.DEFAULTDATA.radionuclides, "numbers", rd.DEFAULTDATA.dataset)
     {'H-3': 1.0}
     >>> H3 = rd.Radionuclide('H-3')
-    >>> rd.inventory._check_dictionary({H3: 1.0}, rd.DEFAULTDATA.radionuclides, "number", rd.DEFAULTDATA.dataset)
+    >>> rd.inventory._check_dictionary({H3: 1.0}, rd.DEFAULTDATA.radionuclides, "numbers", rd.DEFAULTDATA.dataset)
     {'H-3': 1.0}
 
     """
@@ -433,7 +433,7 @@ def _check_dictionary(
                 str(inp) + " is not a valid input for " + str(nuc) + "."
             )
 
-    if input_type != "number":
+    if input_type != "numbers":
         inv_dict = _input_to_number(parsed_inv_dict, input_type, data).copy()
     else:
         inv_dict = parsed_inv_dict.copy()
@@ -494,7 +494,7 @@ class Inventory:
         Dictionary containing radionuclide strings or Radionuclide objects as keys and numbers
         as values.
     input_type : str, optional
-        Type of input (abundance, number, activity; activity is default)
+        Type of input (masses, numbers, activities; activities is default)
     check : bool, optional
         Check for the validity of contents (default is True).
     data : DecayData, optional
@@ -515,7 +515,7 @@ class Inventory:
     >>> H3 = rd.Radionuclide('H-3')
     >>> rd.Inventory({H3: 3.0})
     Inventory: {'H-3': 1682678687.3382246}, decay dataset: icrp107
-    >>> rd.Inventory({'U-238': 21.1, 'Co-57': 7.2}, input_type="mass")
+    >>> rd.Inventory({'U-238': 21.1, 'Co-57': 7.2}, input_type="masses")
     Inventory: {'Co-57': 7.61542657764305e+22, 'U-238': 5.337817684684321e+22}, decay dataset: icrp107
 
     """
@@ -523,7 +523,7 @@ class Inventory:
     def __init__(
         self,
[…]
