+++ tests/test_inventory.py
@@ -56,46 +56,46 @@ class Test(unittest.TestCase):
 
         # Dictionary parsing
         self.assertEqual(
-            _check_dictionary({"H-3": 1.0}, radionuclides, input_type="number"), {"H-3": 1.0}
+            _check_dictionary({"H-3": 1.0}, radionuclides, input_type="numbers"), {"H-3": 1.0}
         )
         self.assertEqual(
-            _check_dictionary({"H3": 1.0}, radionuclides, input_type="number"), {"H-3": 1.0}
+            _check_dictionary({"H3": 1.0}, radionuclides, input_type="numbers"), {"H-3": 1.0}
         )
         self.assertEqual(
-            _check_dictionary({"3H": 1.0}, radionuclides, input_type="number"), {"H-3": 1.0}
+            _check_dictionary({"3H": 1.0}, radionuclides, input_type="numbers"), {"H-3": 1.0}
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1}, radionuclides, input_type="number"), {"H-3": 1}
+            _check_dictionary({"H-3": 1}, radionuclides, input_type="numbers"), {"H-3": 1}
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1}, radionuclides, input_type="number"), {"H-3": 1.0}
+            _check_dictionary({"H-3": 1}, radionuclides, input_type="numbers"), {"H-3": 1.0}
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1.0, "C-14": 2.0}, radionuclides, input_type="number"),
+            _check_dictionary({"H-3": 1.0, "C-14": 2.0}, radionuclides, input_type="numbers"),
             {"H-3": 1.0, "C-14": 2.0},
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1.0, "C-14": 2.0}, radionuclides, input_type="number"),
+            _check_dictionary({"H-3": 1.0, "C-14": 2.0}, radionuclides, input_type="numbers"),
             {"C-14": 2.0, "H-3": 1.0},
         )
         self.assertEqual(
-            _check_dictionary({H3: 1.0, C14: 2.0}, radionuclides, input_type="number"),
+            _check_dictionary({H3: 1.0, C14: 2.0}, radionuclides, input_type="numbers"),
             {"C-14": 2.0, "H-3": 1.0},
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1.0, C14: 2.0}, radionuclides, input_type="number"),
+            _check_dictionary({"H-3": 1.0, C14: 2.0}, radionuclides, input_type="numbers"),
             {"C-14": 2.0, "H-3": 1.0},
         )
         self.assertEqual(
-            _check_dictionary({H3: 1.0, "C-14": 2.0}, radionuclides, input_type="number"),
+            _check_dictionary({H3: 1.0, "C-14": 2.0}, radionuclides, input_type="numbers"),
             {"C-14": 2.0, "H-3": 1.0},
         )
 
         # Catch incorrect arguments
         with self.assertRaises(ValueError):
-            _check_dictionary({"H-3": "1.0"}, radionuclides, input_type="number")
+            _check_dictionary({"H-3": "1.0"}, radionuclides, input_type="numbers")
         with self.assertRaises(ValueError):
-            _check_dictionary({"1.0": "H-3"}, radionuclides, input_type="number")
[…]
