### 2021-07-20 9830a84
Edits to decaydata test outlined in PR #34 review
