### 2021-07-20 69e124b
Ran Black, mypy, pylint
