### 2021-07-21 25f976a
Ran black, pylint, mypy on inventory/test files
