+++ radioactivedecay/inventory.py
@@ -74,9 +74,7 @@ def _add_dictionaries(
 
 
 def _activity_to_number(
-    radionuclide: str,
-    activity : float,
-    data: DecayData = DEFAULTDATA
+    radionuclide: str, activity: float, data: DecayData = DEFAULTDATA
 ) -> float:
     """
     Converts activity to number of atoms.
@@ -105,19 +103,13 @@ def _activity_to_number(
     try:
         index = data.radionuclide_dict[radionuclide]
     except:
-        raise ValueError(
-            radionuclide
-            + " is not a valid member of "
-            + data.dataset
-            )
-        
-    return (activity / data.scipy_data.decay_consts[index])
+        raise ValueError(radionuclide + " is not a valid member of " + data.dataset)
+
+    return activity / data.scipy_data.decay_consts[index]
 
 
 def _mass_to_number(
-    radionuclide: str,
-    mass : float,
-    data: DecayData = DEFAULTDATA
+    radionuclide: str, mass: float, data: DecayData = DEFAULTDATA
 ) -> float:
     """
     Converts a mass to number of atoms.
@@ -146,16 +138,12 @@ def _mass_to_number(
     try:
         index = data.radionuclide_dict[radionuclide]
     except:
-        raise ValueError(
-            radionuclide
-            + " is not a valid member of "
-            + data.dataset
-            )
+        raise ValueError(radionuclide + " is not a valid member of " + data.dataset)
 
-    return (mass / data.masses[index] * Avogadro)
+    return mass / data.masses[index] * Avogadro
 
 
-def _moles_to_number(moles : float) -> float:
+def _moles_to_number(moles: float) -> float:
     """
     Converts number of moles to number of atoms.
 
@@ -180,9 +168,7 @@ def _moles_to_number(moles : float) -> float:
 
 
 def _number_to_activity(
-    radionuclide: str,
-    number : float,
-    data: DecayData = DEFAULTDATA
+    radionuclide: str, number: float, data: DecayData = DEFAULTDATA
 ) -> float:
     """
     Converts number of atoms to activity.
@@ -211,19 +197,13 @@ def _number_to_activity(
     try:
         index = data.radionuclide_dict[radionuclide]
     except:
-        raise ValueError(
-            radionuclide
-            + " is not a valid member of "
-            + data.dataset
-            )
-    
-    return (number * data.scipy_data.decay_consts[index])
+        raise ValueError(radionuclide + " is not a valid member of " + data.dataset)
+
+    return number * data.scipy_data.decay_consts[index]
 
 
 def _number_to_mass(
-    radionuclide: str,
-    number : float,
-    data: DecayData = DEFAULTDATA
+    radionuclide: str, number: float, data: DecayData = DEFAULTDATA
 ) -> float:
     """
     Converts number of atoms to mass in grams.
@@ -252,13 +232,9 @@ def _number_to_mass(
     try:
         index = data.radionuclide_dict[radionuclide]
     except:
-        raise ValueError(
-            radionuclide
-            + " is not a valid member of "
-            + data.dataset
-            )
-    
-    return (number / Avogadro * data.masses[index])
+        raise ValueError(radionuclide + " is not a valid member of " + data.dataset)
+
+    return number / Avogadro * data.masses[index]
 
 
 def _number_to_moles(number: float) -> float:
@@ -281,7 +257,7 @@ def _number_to_moles(number: float) -> float:
     0.49816172015215404
 
     """
-    
+
     return number / Avogadro
 
 
@@ -313,9 +289,7 @@ def _sort_dictionary_alphabetically(
 
 
 def _input_to_number(
-    input_inv_dict: Dict[str, float],
[…]
