### 2021-07-21 a92d1b8
Fix Radionuclide.plot() to cope with stable nuclides in the dataset
