### 2021-07-21 54cee25
Updated README and docs to include API changes
