+++ docs/source/tutorial.rst
@@ -21,19 +21,48 @@ Create an inventory of radionuclides and associated activities as follows:
 
 This is an inventory of natural uranium.
 
-The following commands can be used to view the contents (the radionuclides and
-their activities) and decay data set associated with an inventory:
+The ``Inventory`` class keeps track of the number of atoms
+of each nuclide. The following commands can be used to convert the contents
+(the nuclides and their respective number of atoms) into activities, numbers,
+masses, or abundances, as well as return the decay data set associated with an
+inventory:
 
 .. code-block:: python3
 
-    >>> inv_t0.contents
+    >>> inv_t0.activities()
     {'U-234': 0.005, 'U-235': 0.72, 'U-238': 99.274}
     >>> inv_t0.radionuclides
     ['U-234', 'U-235', 'U-238']
-    >>> inv_t0.activities
-    [0.005, 0.72, 99.274]
+    >>> inv_t0.numbers():
+    {'U-234': 55884417984.51489, 'U-235': 2.3076736283495652e+16, 'U-238': 2.0193793783074922e+19}
+    >>> inv_t0.masses():
+    {'U-234': 2.1718592794623984e-11, 'U-235': 9.006841520909447e-06, 'U-238': 0.00798245788808877}
+    >>> inv_t0.mass_abundances():
+    {'U-234': 2.7177236552303014e-09, 'U-235': 0.0011270576547825488, 'U-238': 0.9988729396274937}
+    >>> inv_t0.moles():
+    {'U-234': 9.279825930955971e-14, 'U-235': 3.831982214161273e-08, 'U-238': 3.3532583491248255e-05}
     >>> inv_t0
-    Inventory: {'U-234': 0.005, 'U-235': 0.72, 'U-238': 99.274}, decay dataset: icrp107
+    Inventory: {'U-234': 55884417984.51489, 'U-235': 2.3076736283495652e+16, 'U-238': 2.0193793783074922e+19}, decay dataset: icrp107
+
+The input of ``Inventory`` is by default activity. The user can easily change
+the input using the ``input_type`` keyword argument during initialization: 
+
+.. code-block:: python3
+
+    # initialize an inventory using activities:
+    >>> inv1 = rd.Inventory({'U-238': 99.274, 'U-235': 0.720, 'U-234': 0.005})
+    >>> inv1.activities()
+    {'U-234': 0.005, 'U-235': 0.72, 'U-238': 99.274}
+    >>> inv1.numbers():
+    {'U-234': 55884417984.51489, 'U-235': 2.3076736283495652e+16, 'U-238': 2.0193793783074922e+19}
+    
+    # initialize an inventory using number of atoms:
+    >>> inv2 = rd.Inventory({'U-238': 2000, 'U-235': 3000, 'U-234': 1500}, input_type="numbers")
+    >>> inv2.activities()
+    {'U-234': 1.3420556696283726e-10, 'U-235': 9.360075764027427e-14, 'U-238': 9.832129719300668e-15}
+    >>> inv2.numbers():
+    {'U-234': 1500, 'U-235': 3000, 'U-238': 2000}
+
 
 Radioactive decay calculations
 ------------------------------
@@ -44,26 +73,26 @@ uranium inventory:
 .. code-block:: python3
 
     >>> inv_t1 = inv_t0.decay(1E9, 'y')
-    >>> inv_t1.contents
+    >>> inv_t1.activities()
[…]
