### 2021-07-22 ffc3d5b
Add references for AMDC data, decay chain images now show explicitly if nuclides are stable, update contributors and acknowledgements
