+++ docs/source/overview.rst
@@ -10,10 +10,12 @@ decay calculations, and output decay data for radionuclides and decay chains.
 
 The original goal was to create a light-weight Python package for radioactive
 decay calculations, with full support for branching decays, multi-step decay
-chains, and metastable states. ``radioactivedecay`` uses the decay data from
-ICRP Publication 107 :ref:`[1] <refs>` by default. It solves the radioactive
+chains, and metastable states. By default ``radioactivedecay`` uses decay data
+from ICRP Publication 107 :ref:`[1] <refs>` and atomic mass data from the
+Atomic Mass Data Center (AMDC AME2020 and Nubase 2020 evaluations)
+:ref:`[2] <refs>`, :ref:`[3] <refs>` and :ref:`[4] <refs>`. It solves the radioactive
 decay differential equations analytically using basic linear algebra operations
-:ref:`[2] <refs>`.
+:ref:`[5] <refs>`.
 
 In order to use ``radioactivedecay``, you will need Python 3.6+ with the
 Matplotlib, NetworkX, NumPy, SciPy and SymPy packages installed. The code is
@@ -91,11 +93,11 @@ How it works
 
 ``radioactivedecay`` calculates an analytical solution to the decay chain
 differential equations using matrix and vector multiplications. It implements
-the method described in ref. :ref:`[2] <refs>`.  See the
+the method described in ref. :ref:`[5] <refs>`.  See the
 `theory docpage <https://alexmalins.com/radioactivedecay/theory.html>`_ for
-more details. It calls NumPy :ref:`[3] <refs>` and SciPy :ref:`[4] <refs>` for
+more details. It calls NumPy :ref:`[6] <refs>` and SciPy :ref:`[7] <refs>` for
 the matrix operations. There is also a high numerical precision decay
-calculation mode based on SymPy :ref:`[5] <refs>` routines.
+calculation mode based on SymPy :ref:`[8] <refs>` routines.
 
 The `notebooks directory 
 <https://github.com/alexmalins/radioactivedecay/tree/main/notebooks>`_ 
@@ -104,9 +106,9 @@ in the GitHub repository contains some Jupyter Notebooks for creating the
 <https://github.com/alexmalins/radioactivedecay/tree/main/notebooks/icrp107_dataset/icrp107_dataset.ipynb>`_
 for ``radioactivedecay``, and cross-checks against `PyNE
 <https://github.com/alexmalins/radioactivedecay/tree/main/notebooks/comparisons/pyne/rd_pyne_truncated_compare.ipynb>`_ 
-:ref:`[6] <refs>` and `Radiological Toolbox 
+:ref:`[9] <refs>` and `Radiological Toolbox 
 <https://github.com/alexmalins/radioactivedecay/tree/main/notebooks/comparisons/radiological_toolbox/radiological_toolbox_compare.ipynb>`_
-:ref:`[7] <refs>`.
+:ref:`[10] <refs>`.
 
 Limitations
 -----------
@@ -139,7 +141,7 @@ There are also some limitations associated with the ICRP-107 decay dataset:
 * There are a few instances where minor decay pathways were not included in
   ICRP-107, e.g. the decay pathways for At-219-> Rn-219 (|beta| ~3%), Es-250 ->
[…]
