### 2021-07-23 4739062
Added mole_fractions(), edits outlined in PR #35
