+++ radioactivedecay/inventory.py
@@ -102,8 +102,9 @@ def _activity_to_number(
 
     try:
         index = data.radionuclide_dict[radionuclide]
-    except:
-        raise ValueError(radionuclide + " is not a valid member of " + data.dataset)
+    except BaseException:
+        raise ValueError(radionuclide + " is not a valid member of "
+                         + data.dataset)
 
     return activity / data.scipy_data.decay_consts[index]
 
@@ -137,8 +138,9 @@ def _mass_to_number(
 
     try:
         index = data.radionuclide_dict[radionuclide]
-    except:
-        raise ValueError(radionuclide + " is not a valid member of " + data.dataset)
+    except BaseException:
+        raise ValueError(radionuclide + " is not a valid member of "
+                         + data.dataset)
 
     return mass / data.masses[index] * Avogadro
 
@@ -196,8 +198,9 @@ def _number_to_activity(
 
     try:
         index = data.radionuclide_dict[radionuclide]
-    except:
-        raise ValueError(radionuclide + " is not a valid member of " + data.dataset)
+    except BaseException:
+        raise ValueError(radionuclide + " is not a valid member of "
+                         + data.dataset)
 
     return number * data.scipy_data.decay_consts[index]
 
@@ -231,8 +234,9 @@ def _number_to_mass(
 
     try:
         index = data.radionuclide_dict[radionuclide]
-    except:
-        raise ValueError(radionuclide + " is not a valid member of " + data.dataset)
+    except BaseException:
+        raise ValueError(
+            radionuclide + " is not a valid member of " + data.dataset)
 
     return number / Avogadro * data.masses[index]
 
@@ -303,7 +307,6 @@ def _input_to_number(
         Format of the input.
     data : DecayData
         Decay dataset (default is the ICRP-107 dataset).
-    
 
     Returns
     -------
@@ -400,7 +403,8 @@ def _check_dictionary(
     }
     for nuc, inp in parsed_inv_dict.items():
         if not isinstance(inp, (float, int)):
-            raise ValueError(str(inp) + " is not a valid input for " + str(nuc) + ".")
+            raise ValueError(
+                str(inp) + " is not a valid input for " + str(nuc) + ".")
 
     if input_type != "numbers":
         inv_dict = _input_to_number(parsed_inv_dict, input_type, data).copy()
@@ -472,7 +476,7 @@ class Inventory:
     Attributes
     ----------
     numbers : dict
-        Dictionary containing radionuclide strings as keys and numbers as values. Radionuclides
+        Dictionary containing nuclide string as keys and number of nuclides as values. Radionuclides
         in contents are sorted alphabetically.
     data : DecayData
         Decay dataset.
@@ -480,12 +484,12 @@ class Inventory:
     Examples
     --------
     >>> rd.Inventory({'Tc-99m': 2.3, 'I-123': 5.8})
-    Inventory: {'I-123': 399738.47946141585, 'Tc-99m': 71852.27235544211}, decay dataset: icrp107
+    Inventory activities: {'I-123': 2.3, 'Tc-99m': 5.8}, decay dataset: icrp107
     >>> H3 = rd.Radionuclide('H-3')
     >>> rd.Inventory({H3: 3.0})
-    Inventory: {'H-3': 1682678687.3382246}, decay dataset: icrp107
+    Inventory activities: {'H-3': 3.0}, decay dataset: icrp107
     >>> rd.Inventory({'U-238': 21.1, 'Co-57': 7.2}, input_type="masses")
-    Inventory: {'Co-57': 7.61542657764305e+22, 'U-238': 5.337817684684321e+22}, decay dataset: icrp107
[…]
