+++ CONTRIBUTING.md
@@ -1,44 +1,105 @@
 # Contributing
 
-Users are welcome to submit bug reports, feature requests, suggestions for
-improvements and pull requests for this package. Below are some guidelines
-to keep in mind when requesting or making changes.
+Anybody is welcome to submit bug reports and make feature requests. You can
+also submit pull requests (PRs) with new code to improve this package. Below
+are some general guidelines for requesting changes or submitting a PR.
 
 
 ## Reporting Issues
 
-If you encounter an issue or find a bug in the code, please open an issue on
-the GitHub [issues](https://github.com/alexmalins/radioactivedecay/issues)
-page.
+If you encounter an issue or find a bug in this package, please open an issue
+on the
+[GitHub issues page](https://github.com/alexmalins/radioactivedecay/issues).
 
-When submitting issues please include the version of ``radioactivedecay`` and
-Python you are using, and include a minimal example of code or information
-which demonstrates the problem.
-
-
-## Making Changes
-
-If you would like to contribute code, bug fixes or documentation to the
-project, please fork the GitHub repository, make your changes, then submit a
-[pull request](https://github.com/alexmalins/radioactivedecay/pulls).
+When submitting an issue please state the version of ``radioactivedecay`` that
+you are using, and include a minimal example of code or information which
+demonstrates the problem.
 
 
-## Docstrings
+## Pull Requests
 
-Docstrings use the [NumPy format](https://numpydoc.readthedocs.io/en/latest/format.html).
-The doctrings are automatically harvested by Sphinx to create the API section
-of the [documentation](https://alexmalins.com/radioactivedecay/).
+If you would like to contribute code, bug fixes or documentation improvements
+to the project, please fork the GitHub repository, make your changes, then
+submit a [pull request](https://github.com/alexmalins/radioactivedecay/pulls).
+
+Below are the code standards used by this project. Note they are not rigid
+requirements, but goals to aim for when submitting a PR. If you are not sure
+how to do something, please feel free to contact one of the existing developers
+for assistance.
 
 
-## Release guidelines
+### Code Formatting & Linting
 
-To release a new version of ``radioactivedecay``:
-* Update version number in setup.py
-* Update version number in radioactivedecay/__init__.py
-* Update version number in docs/source/conf.py
-* Run the tests: ``coverage run -m unittest discover``
-* Upload coverage reports to codecov
-* Compile docs: ``docs/make html``
-* Upload docs to https://alexmalins.com/radioactivedecay/
-* Create a release on GitHub
-* Upload new version to PyPi
+The project uses [Black](https://black.readthedocs.io/en/stable/) to
+standardize code formatting. Black is a command line utility that can be
+installed via [pip](https://pypi.org/project/black/) or
+[conda-forge](https://anaconda.org/conda-forge/black).
+
+Please run ``black .`` from the base directory before making commits for PRs.
+This will automatically format the code into a standard format.
+
+[Pylint](https://www.pylint.org/) is used for basic error checking and
+refactoring (installable from [pip](https://pypi.org/project/pylint/) or
+[conda-forge](https://anaconda.org/conda-forge/pylint)). Run
+``pylint radioactivedecay`` from the base directory and try to fix any new
+warnings that are raised.
+
+
+### Type Hints
+
+This project requires Python 3.6+ and type hints are included in the headers of
+all functions and methods exposed to users. Please add type hints for any new
+functions you create.
+
+Type hints are checked using [Mypy](http://mypy-lang.org/). Mypy is installable
+via [pip](https://pypi.org/project/mypy/) or
+[conda-forge](https://anaconda.org/conda-forge/mypy).
[…]
