### 2021-07-28 22e74b8
CONTRIBUTING.md & README.md minor fixes, add docstring example to radionuclide.py
