+++ README.md
@@ -81,30 +81,29 @@ Additional options for inputs and outputs include masses and numbers of atoms,
 using the ``input_type`` argument and ``numbers()``, ``masses()``,
 ``mass_abundances()``, and ``moles()`` methods. 
 
-.. code-block:: python3
+```pycon
+>>> inv_mass_t0 = rd.Inventory({'H-3': 3.2}, input_type="masses")
+>>> inv_mass_t1 = inv_mass_t0.decay(12.32, 'y')
+>>> inv_mass_t1.masses()
+{'H-3': 1.6000000000000003, 'He-3': 1.5999894116584246}
 
-    >>> inv_mass_t0 = rd.Inventory({'H-3': 3.2}, input_type="masses")
-    >>> inv_mass_t1 = inv_mass_t0.decay(12.32, 'y')
-    >>> inv_mass_t1.masses()
-    {'H-3': 1.6000000000000003, 'He-3': 1.5999894116584246}
-
-    >>> inv_num_t0 = rd.Inventory({'C-14': 3.2E24}, input_type="numbers")
-    >>> inv_num_t1 = inv_num_t0.decay(3000, 'y')
-    >>> inv_num_t1.moles()
-    {'C-14': 3.6894551567795797, 'N-14': 1.6242698581767292}
+>>> inv_num_t0 = rd.Inventory({'C-14': 3.2E24}, input_type="numbers")
+>>> inv_num_t1 = inv_num_t0.decay(3000, 'y')
+>>> inv_num_t1.moles()
+{'C-14': 3.6894551567795797, 'N-14': 1.6242698581767292}
+```
 
 Mass follows slightly stricter rules for units: mass in grams, with the added
 ability to input mass abundances using an inventory with masses summing to 1.0:
 
-.. code-block:: python3
-
-    >>> inv_abund_t0 = rd.Inventory({'Ni-56': .8, 'Co-56': .2}, input_type="masses")
-    >>> inv_abund_t1 = inv_abund_t0.decay(35.0, 'd')
-    >>> inv_abund_t1.mass_abundances()
-    {'Co-56': 0.7643201942234104,
-     'Fe-56': 0.2209301066281599,
-     'Ni-56': 0.014749699148429682}
-
+```pycon
+>>> inv_abund_t0 = rd.Inventory({'Ni-56': .8, 'Co-56': .2}, input_type="masses")
+>>> inv_abund_t1 = inv_abund_t0.decay(35.0, 'd')
+>>> inv_abund_t1.mass_abundances()
+{'Co-56': 0.7643201942234104,
+ 'Fe-56': 0.2209301066281599,
+  'Ni-56': 0.014749699148429682}
+```
 
 ### Plotting decay graphs
 
