### 2021-07-28 1f71663
Refactoring
- Update version numbers
- Prefer nuclide over radionuclide for variable names as now datasets contain stable nuclides
- Change DecayData.dataset attribute to DecayData.dataset_name (more explicit variable naming)
- Fix Inventory.decay() new dictionary creation bug
- Fix Inventory.__repr__() test
- Code formatting with latest version of Black (21.7b0)
- Fix some pylint warnings
- Modify Inventory quantity conversion functions not to take in duplicate data from a decay dataset
- Remove default parameter settings from functions that are not exposed to users
