+++ docs/source/conf.py
@@ -24,7 +24,7 @@ copyright = "2021, Alex Malins"
 author = "Alex Malins"
 
 # The full version, including alpha/beta/rc tags
-release = "0.3.4"
+release = "0.4.0"
 
 
 # -- General configuration ---------------------------------------------------
