+++ tests/test_inventory.py
@@ -34,6 +34,9 @@ class Test(unittest.TestCase):
             {"Pm-141": 4.0, "Rb-78": 2.0, "Rb-90": 4.0},
         )
 
+    def test__activity_to_number(self):
+        pass
+
     def test__sort_dictionary_alphabetically(self):
         """
         Test the sorting of a dictionary by its keys alphabetically.
@@ -50,65 +53,67 @@ class Test(unittest.TestCase):
         Test the checking of inventory dictionaries.
         """
 
-        radionuclides = ["H-3", "C-14"]
+        nuclides = ["H-3", "C-14"]
         H3 = Radionuclide("H3")
         C14 = Radionuclide("C14")
 
         # Dictionary parsing
         self.assertEqual(
-            _check_dictionary({"H-3": 1.0}, radionuclides, input_type="numbers"),
+            _check_dictionary({"H-3": 1.0}, input_type="numbers", data=DEFAULTDATA),
             {"H-3": 1.0},
         )
         self.assertEqual(
-            _check_dictionary({"H3": 1.0}, radionuclides, input_type="numbers"),
+            _check_dictionary({"H3": 1.0}, input_type="numbers", data=DEFAULTDATA),
             {"H-3": 1.0},
         )
         self.assertEqual(
-            _check_dictionary({"3H": 1.0}, radionuclides, input_type="numbers"),
+            _check_dictionary({"3H": 1.0}, input_type="numbers", data=DEFAULTDATA),
             {"H-3": 1.0},
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1}, radionuclides, input_type="numbers"),
+            _check_dictionary({"H-3": 1}, input_type="numbers", data=DEFAULTDATA),
             {"H-3": 1},
         )
         self.assertEqual(
-            _check_dictionary({"H-3": 1}, radionuclides, input_type="numbers"),
+            _check_dictionary({"H-3": 1}, input_type="numbers", data=DEFAULTDATA),
             {"H-3": 1.0},
         )
         self.assertEqual(
             _check_dictionary(
-                {"H-3": 1.0, "C-14": 2.0}, radionuclides, input_type="numbers"
+                {"H-3": 1.0, "C-14": 2.0}, input_type="numbers", data=DEFAULTDATA
             ),
             {"H-3": 1.0, "C-14": 2.0},
         )
         self.assertEqual(
             _check_dictionary(
-                {"H-3": 1.0, "C-14": 2.0}, radionuclides, input_type="numbers"
-            ),
-            {"C-14": 2.0, "H-3": 1.0},
-        )
-        self.assertEqual(
-            _check_dictionary({H3: 1.0, C14: 2.0}, radionuclides, input_type="numbers"),
-            {"C-14": 2.0, "H-3": 1.0},
-        )
-        self.assertEqual(
-            _check_dictionary(
-                {"H-3": 1.0, C14: 2.0}, radionuclides, input_type="numbers"
+                {"H-3": 1.0, "C-14": 2.0}, input_type="numbers", data=DEFAULTDATA
             ),
             {"C-14": 2.0, "H-3": 1.0},
         )
         self.assertEqual(
             _check_dictionary(
-                {H3: 1.0, "C-14": 2.0}, radionuclides, input_type="numbers"
+                {H3: 1.0, C14: 2.0}, input_type="numbers", data=DEFAULTDATA
+            ),
+            {"C-14": 2.0, "H-3": 1.0},
+        )
+        self.assertEqual(
+            _check_dictionary(
[…]
