+++ radioactivedecay/decaydata.py
@@ -25,7 +25,7 @@ from scipy import sparse
 from sympy import log, Matrix
 from sympy.core.numbers import Rational
 from sympy.matrices import SparseMatrix
-from radioactivedecay.utils import parse_nuclide, parse_radionuclide, time_unit_conv
+from radioactivedecay.utils import parse_nuclide, time_unit_conv
 
 try:
     from importlib import resources
@@ -33,14 +33,14 @@ except ImportError:
     import importlib_resources as resources
 
 
-def _get_package_filepath(dataset: str, filename: str) -> ContextManager[Path]:
+def _get_package_filepath(dataset_name: str, filename: str) -> ContextManager[Path]:
     """
     Returns the path to a decay dataset file which is bundled as a sub-package within the
     ``radioactivedecay`` package.
 
     Parameters
     ----------
-    dataset : str
+    dataset_name : str
         Name of the decay dataset.
     filename : str
         Name of the file.
@@ -51,18 +51,20 @@ def _get_package_filepath(dataset: str, filename: str) -> ContextManager[Path]:
         A context manager providing a file path object for the decay dataset file.
     """
 
-    with resources.path(__package__ + "." + dataset, filename) as package_path:
+    with resources.path(__package__ + "." + dataset_name, filename) as package_path:
         return package_path
 
 
-def _get_package_pickle(dataset: str, filename: str) -> Union[Matrix, SparseMatrix]:
+def _get_package_pickle(
+    dataset_name: str, filename: str
+) -> Union[Matrix, SparseMatrix]:
     """
     Returns an object loaded from a decay dataset file which is bundled as a sub-package within the
     ``radioactivedecay`` package.
 
     Parameters
     ----------
-    dataset : str
+    dataset_name : str
         Name of the decay dataset.
     filename : str
         Name of the file.
@@ -73,7 +75,7 @@ def _get_package_pickle(dataset: str, filename: str) -> Union[Matrix, SparseMatr
         A SymPy dense or sparse matrix loaded from the decay dataset pickle file.
     """
 
-    with resources.open_binary(__package__ + "." + dataset, filename) as file:
+    with resources.open_binary(__package__ + "." + dataset_name, filename) as file:
         return pickle.load(file)
 
 
@@ -134,7 +136,7 @@ class DecayMatrices:
         The matrix exponential that is used in radioactive decay calculations. It is a diagonal
         matrix that is pre-allocted for performance reasons.
     vector_n0 : numpy.ndarray or sympy.matrices.dense.MutableDenseMatrix
-        Column vector for the number of atoms of each radionuclide. It is pre-allocted for
+        Column vector for the number of atoms of each nuclide. It is pre-allocted for
         performance reasons.
     year_conv : float or sympy.core.numbers.Rational
         Conversion factor for number of days in one year.
@@ -217,7 +219,7 @@ class DecayData:
 
     Parameters
     ----------
-    dataset : str
+    dataset_name : str
         Name of the decay dataset.
     dir_path : str or None, optional
         Path to the directory containing the decay dataset files. Use None if the data are bundled
@@ -228,17 +230,19 @@ class DecayData:
 
     Attributes
     ----------
-    dataset : str
+    dataset_name : str
         Name of the decay dataset.
     hldata : numpy.ndarray
         List of tuples containing half-life floats, time unit strings and readable format half-life
         strings.
-    num_radionuclides : int
-        Number of radionuclides in the dataset.
-    radionuclides : numpy.ndarray
-        NumPy array of radionuclides in the dataset (string format is 'H-3', etc.).
-    radionuclide_dict : dict
[…]
