+++ tests/test_radionuclide.py
@@ -17,7 +17,7 @@ class Test(unittest.TestCase):
         """
 
         nuc = Radionuclide("Rn-222")
-        self.assertEqual(nuc.radionuclide, "Rn-222")
+        self.assertEqual(nuc.nuclide, "Rn-222")
         self.assertEqual(nuc.prog_bf_mode, {"Po-218": [1.0, "\u03b1"]})
 
     def test_radionuclide_half_life(self):
@@ -116,7 +116,7 @@ class Test(unittest.TestCase):
 
         nuc = Radionuclide("K-40")
         data = DecayData("icrp107", load_sympy=True)
-        self.assertEqual(hash(nuc), hash(("K-40", data.dataset)))
+        self.assertEqual(hash(nuc), hash(("K-40", data.dataset_name)))
 
 
 if __name__ == "__main__":

[…]
