### 2021-07-28 2974ee0
Updates for new Inventory functions
- Remove checks of whether nuclide is already in the dataset from Inventory quantity conversion functions (this is already checked in _check_dictionary() via parse_nuclide() calls or by virtue of the nuclide already being a member of the Inventory).
- Add tests for Inventory quantity conversion functions
