### 2021-07-28 c7b5269
Add unit conversion utils for more quantities, run black
