+++ radioactivedecay/utils.py
@@ -258,3 +258,352 @@ def time_unit_conv_sympy(
         )
 
     return time_period * conv[units_from] / conv[units_to]
+
+
+def activity_unit_conv(activity: float, units_from: str, units_to: str) -> float:
+    """
+    Converts an activity from one unit to another.
+
+    Parameters
+    ----------
+    activity : float
+        Activity before conversion.
+    units_from : str
+        Activity unit before conversion
+    units_to : str
+        Activity unit after conversion
+
+    Returns
+    -------
+    float
+        Activity in new units.
+
+    Raises
+    ------
+    ValueError
+        If one of the activity unit parameters is invalid.
+
+    Examples
+    --------
+    >>> rd.utils.activity_unit_conv(1.0, 'Ci', 'Bq')
+    3.7e10
+
+    """
+
+    curie_to_Bq = 3.7e10
+
+    conv = {
+        "pBq": 1.0e-12,
+        "nBq": 1.0e-9,
+        "μBq": 1.0e-6,
+        "uBq": 1.0e-6,
+        "mBq": 1.0e-3,
+        "Bq": 1.0,
+        "kBq": 1.0e3,
+        "MBq": 1.0e6,
+        "GBq": 1.0e9,
+        "TBq": 1.0e12,
+        "PBq": 1.0e15,
+        "EBq": 1.0e18,
+        "pCi": 1.0e-12 * curie_to_Bq,
+        "nCi": 1.0e-9 * curie_to_Bq,
+        "μCi": 1.0e-6 * curie_to_Bq,
+        "uCi": 1.0e-6 * curie_to_Bq,
+        "mCi": 1.0e-3 * curie_to_Bq,
+        "Ci": 1.0 * curie_to_Bq,
+        "kCi": 1.0e3 * curie_to_Bq,
+        "MCi": 1.0e6 * curie_to_Bq,
+        "GCi": 1.0e9 * curie_to_Bq,
+        "TCi": 1.0e12 * curie_to_Bq,
+        "PCi": 1.0e15 * curie_to_Bq,
+        "ECi": 1.0e18 * curie_to_Bq,
+        "dpm": 60.0,
+    }
+
+    if units_from not in conv:
+        raise ValueError(
+            str(units_from)
+            + ' is not a valid activitiy unit, e.g. "Bq", "kBq", "Ci"...'
+        )
+    if units_to not in conv:
+        raise ValueError(
+            str(units_to) + ' is not a activitiy unit, e.g. "Bq", "kBq", "Ci"...'
+        )
+
+    return activity * conv[units_from] / conv[units_to]
+
+
+def activity_unit_conv_sympy(
+    activity: Rational, units_from: str, units_to: str
+) -> Rational:
+    """
+    Same functionality as activity_unit_conv(), but uses SymPy arbitrary-precision arithmetic.
+
+    Parameters
+    ----------
+    activity : sympy.core.numbers.Rational
+        Activity before conversion.
+    units_from : str
+        Activity unit before conversion
+    units_to : str
+        Activity unit after conversion
+
+    Returns
+    -------
+    sympy.core.numbers.Rational
+        Activity in new units.
+
+    Raises
+    ------
+    ValueError
+        If one of the activity unit parameters is invalid.
+
+    """
+
+    curie_to_Bq = 37000000000
+
+    conv = {
+        "pBq": Integer(1) / 1000000000000,
+        "nBq": Integer(1) / 1000000000,
+        "μBq": Integer(1) / 1000000,
+        "uBq": Integer(1) / 1000000,
+        "mBq": Integer(1) / 1000,
+        "Bq": Integer(1),
+        "kBq": Integer(1000),
+        "MBq": Integer(1000000),
[…]
