+++ tests/test_utils.py
@@ -9,6 +9,12 @@ from radioactivedecay.utils import (
     parse_nuclide,
     time_unit_conv,
     time_unit_conv_sympy,
+    activity_unit_conv,
+    activity_unit_conv_sympy,
+    mass_unit_conv,
+    mass_unit_conv_sympy,
+    moles_unit_conv,
+    moles_unit_conv_sympy,
 )
 
 
@@ -320,6 +326,278 @@ class Test(unittest.TestCase):
             time_unit_conv_sympy(1.0, "y", "ty", yconv)
         with self.assertRaises(ValueError):
             time_unit_conv_sympy(1.0, "ty", 1.0, yconv)
+
+    def test_activity_unit_conv(self):
+        """
+        Test the conversion between activity units.
+        """
+
+        self.assertEqual(activity_unit_conv(1.0, "Bq", "Bq"), 1.0e0)
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "pBq"), 1.0e12, places=(15 - 12)
+        )
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "nBq"), 1.0e9, places=(15 - 9)
+        )
+        self.assertEqual(activity_unit_conv(1.0, "Bq", "μBq"), 1.0e6)
+        self.assertEqual(activity_unit_conv(1.0, "Bq", "uBq"), 1.0e6)
+        self.assertEqual(activity_unit_conv(1.0, "Bq", "mBq"), 1.0e3)
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "kBq"), 1.0e-3, places=(3 + 15)
+        )
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "MBq"), 1.0e-6, places=(6 + 15)
+        )
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "GBq"), 1.0e-9, places=(9 + 15)
+        )
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "TBq"), 1.0e-12, places=(12 + 15)
+        )
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "PBq"), 1.0e-15, places=(15 + 15)
+        )
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "EBq"), 1.0e-18, places=(18 + 15)
+        )
+
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "pCi"), 1.0e12 / 3.7e10, places=(15 - 12)
+        )
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "nCi"), 1.0e9 / 3.7e10, places=(15 - 9)
+        )
+        self.assertEqual(activity_unit_conv(1.0, "Bq", "μCi"), 1.0e6 / 3.7e10)
+        self.assertEqual(activity_unit_conv(1.0, "Bq", "uCi"), 1.0e6 / 3.7e10)
+        self.assertEqual(activity_unit_conv(1.0, "Bq", "mCi"), 1.0e3 / 3.7e10)
+        self.assertEqual(activity_unit_conv(1.0, "Bq", "Ci"), 1.0 / 3.7e10)
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "kCi"), 1.0e-3 / 3.7e10, places=(3 + 15)
+        )
+        self.assertAlmostEqual(
+            activity_unit_conv(1.0, "Bq", "MCi"), 1.0e-6 / 3.7e10, places=(6 + 15)
+        )
[…]
