+++ radioactivedecay/inventory.py
@@ -25,6 +25,12 @@ from radioactivedecay.decaydata import DecayData, DEFAULTDATA, np
 from radioactivedecay.plots import _decay_graph
 from radioactivedecay.radionuclide import Radionuclide
 from radioactivedecay.utils import (
+    activity_units,
+    activity_unit_conv,
+    mass_units,
+    mass_unit_conv,
+    moles_units,
+    moles_unit_conv,
     parse_nuclide,
     time_unit_conv,
     time_unit_conv_sympy,
@@ -282,7 +288,7 @@ def _sort_dictionary_alphabetically(
 
 
 def _input_to_number(
-    input_inv_dict: Dict[str, float], input_type: str, data: DecayData
+    input_inv_dict: Dict[str, float], units: str, data: DecayData
 ) -> Dict[str, float]:
     """
     Converts mass, moles, or activity input inventory dictionary into number.
@@ -290,10 +296,10 @@ def _input_to_number(
     Parameters
     ----------
     input_inv_dict : dict
-        Dictionary containing nuclide strings or Radionuclide objects as keys and masses,
+        Dictionary containing nuclide strings or Radionuclide objects as keys, and masses,
         moles, or activities as values.
-    input_type : str
-        The input type ('activities', 'moles' or 'masses').
+    units : str
+        Units of the values in the input dictionary.
     data : DecayData
         Decay dataset.
 
@@ -318,30 +324,32 @@ def _input_to_number(
 
     number_dict = {}
 
-    if input_type == "activities":
+    if units in activity_units:
         for nuc, act in input_inv_dict.items():
-            number = _activity_to_number(nuc, act, data)
+            number = _activity_to_number(
+                nuc, activity_unit_conv(act, units, "Bq"), data
+            )
             number_dict[nuc] = number
         converted_dict = number_dict.copy()
-    elif input_type == "moles":
+    elif units in moles_units:
         for nuc, moles in input_inv_dict.items():
-            number = _moles_to_number(moles)
+            number = _moles_to_number(moles_unit_conv(moles, units, "mol"))
             number_dict[nuc] = number
         converted_dict = number_dict.copy()
-    elif input_type == "masses":
+    elif units in mass_units:
         for nuc, mass in input_inv_dict.items():
-            number = _mass_to_number(nuc, mass, data)
+            number = _mass_to_number(nuc, mass_unit_conv(mass, units, "g"), data)
             number_dict[nuc] = number
         converted_dict = number_dict.copy()
     else:
-        raise ValueError(input_type + " is not a valid value input type")
+        raise ValueError(units + " is not a valid value input type")
 
     return converted_dict
 
 
 def _check_dictionary(
     input_inv_dict: Dict[Union[str, Radionuclide], float],
-    input_type: str,
+    units: str,
     data: DecayData,
 ) -> Dict[str, float]:
     """
@@ -353,8 +361,8 @@ def _check_dictionary(
     input_inv_dict : dict
         Dictionary containing nuclide strings or Radionuclide objects as keys and quantities
         as values.
-    input_type : str
-        The input type ('activities', 'moles' or 'masses').
+    units : str
+        Units of the values in the input dictionary (e.g. 'Bq', 'g', 'mol', 'num').
     data : DecayData
         Decay dataset.
 
@@ -393,8 +401,8 @@ def _check_dictionary(
                 str(inp) + " is not a valid amount of nuclide " + str(nuc) + "."
             )
 
-    if input_type != "numbers":
-        inv_dict = _input_to_number(parsed_inv_dict, input_type, data).copy()
+    if units != "num":
+        inv_dict = _input_to_number(parsed_inv_dict, units, data).copy()
     else:
[…]
