+++ radioactivedecay/utils.py
@@ -260,6 +260,35 @@ def time_unit_conv_sympy(
     return time_period * conv[units_from] / conv[units_to]
 
 
+activity_units = {
+    "pBq": 1.0e-12,
+    "nBq": 1.0e-9,
+    "μBq": 1.0e-6,
+    "uBq": 1.0e-6,
+    "mBq": 1.0e-3,
+    "Bq": 1.0,
+    "kBq": 1.0e3,
+    "MBq": 1.0e6,
+    "GBq": 1.0e9,
+    "TBq": 1.0e12,
+    "PBq": 1.0e15,
+    "EBq": 1.0e18,
+    "pCi": 1.0e-12 * 3.7e10,
+    "nCi": 1.0e-9 * 3.7e10,
+    "μCi": 1.0e-6 * 3.7e10,
+    "uCi": 1.0e-6 * 3.7e10,
+    "mCi": 1.0e-3 * 3.7e10,
+    "Ci": 1.0 * 3.7e10,
+    "kCi": 1.0e3 * 3.7e10,
+    "MCi": 1.0e6 * 3.7e10,
+    "GCi": 1.0e9 * 3.7e10,
+    "TCi": 1.0e12 * 3.7e10,
+    "PCi": 1.0e15 * 3.7e10,
+    "ECi": 1.0e18 * 3.7e10,
+    "dpm": 60.0,
+}
+
+
 def activity_unit_conv(activity: float, units_from: str, units_to: str) -> float:
     """
     Converts an activity from one unit to another.
@@ -290,47 +319,17 @@ def activity_unit_conv(activity: float, units_from: str, units_to: str) -> float
 
     """
 
-    curie_to_Bq = 3.7e10
-
-    conv = {
-        "pBq": 1.0e-12,
-        "nBq": 1.0e-9,
-        "μBq": 1.0e-6,
-        "uBq": 1.0e-6,
-        "mBq": 1.0e-3,
-        "Bq": 1.0,
-        "kBq": 1.0e3,
-        "MBq": 1.0e6,
-        "GBq": 1.0e9,
-        "TBq": 1.0e12,
-        "PBq": 1.0e15,
-        "EBq": 1.0e18,
-        "pCi": 1.0e-12 * curie_to_Bq,
-        "nCi": 1.0e-9 * curie_to_Bq,
-        "μCi": 1.0e-6 * curie_to_Bq,
-        "uCi": 1.0e-6 * curie_to_Bq,
-        "mCi": 1.0e-3 * curie_to_Bq,
-        "Ci": 1.0 * curie_to_Bq,
-        "kCi": 1.0e3 * curie_to_Bq,
-        "MCi": 1.0e6 * curie_to_Bq,
-        "GCi": 1.0e9 * curie_to_Bq,
-        "TCi": 1.0e12 * curie_to_Bq,
-        "PCi": 1.0e15 * curie_to_Bq,
-        "ECi": 1.0e18 * curie_to_Bq,
-        "dpm": 60.0,
-    }
-
-    if units_from not in conv:
+    if units_from not in activity_units:
         raise ValueError(
             str(units_from)
             + ' is not a valid activitiy unit, e.g. "Bq", "kBq", "Ci"...'
         )
-    if units_to not in conv:
+    if units_to not in activity_units:
         raise ValueError(
             str(units_to) + ' is not a activitiy unit, e.g. "Bq", "kBq", "Ci"...'
         )
 
-    return activity * conv[units_from] / conv[units_to]
+    return activity * activity_units[units_from] / activity_units[units_to]
 
 
 def activity_unit_conv_sympy(
@@ -360,8 +359,6 @@ def activity_unit_conv_sympy(
 
     """
 
-    curie_to_Bq = 37000000000
[…]
