### 2021-07-30 a3d4b13
Add tests for new inventory functions and methods, fix Inventory.plot() yunits
