+++ radioactivedecay/inventory.py
@@ -1,10 +1,10 @@
 """
 The inventory module defines the ``Inventory`` class. Each ``Inventory`` instance contains one or
-more nuclide each with an associated number of atoms. The decay of the nuclide(s) in an
+more nuclide each with an associated quantity (number of atoms). The decay of the nuclide(s) in an
 ``Inventory`` can be calculated by using the ``decay()`` method (normal double-precision
 floating-point operations). Use the ``decay_high_precision()`` method to perform a SymPy high
-numerical precision calculation. A ``DecayData`` dataset associated with each ``Inventory``
-instance is the radioactive decay data source (default is rd.DEFAULTDATA).
+numerical precision calculation. A ``DecayData`` dataset is associated with each ``Inventory``
+instance (default is rd.DEFAULTDATA).
 
 The code examples shown in the docstrings assume the ``radioactivedecay`` package has been imported
 as:
@@ -173,7 +173,7 @@ def _moles_to_number(moles: float) -> float:
 
 def _number_to_activity(nuclide: str, number: float, data: DecayData) -> float:
     """
-    Converts number of atoms to activity.
+    Converts number of atoms to activity in Bq.
 
     Parameters
     ----------
@@ -187,7 +187,7 @@ def _number_to_activity(nuclide: str, number: float, data: DecayData) -> float:
     Returns
     -------
     float
-        Activity of the nuclide.
+        Activity of the nuclide in Bq.
 
     Raises
     ------
@@ -291,7 +291,8 @@ def _input_to_number(
     input_inv_dict: Dict[str, float], units: str, data: DecayData
 ) -> Dict[str, float]:
     """
-    Converts mass, moles, or activity input inventory dictionary into number.
+    Converts an inventory dictionary where the values are mass, moles, or activity into one where
+    the values are number of atoms.
 
     Parameters
     ----------
@@ -315,36 +316,32 @@ def _input_to_number(
 
     Examples
     --------
-    >>> rd.inventory._input_to_number({'Ni-56': .3, 'Co-56': .7}, 'activities', rd.DEFAULTDATA)
-    {'Ni-56': 227172.53191853972, 'Co-56': 6738641.562715049}
-    >>> rd.inventory._input_to_number({'U-238': 21.1, 'Co-57': 7.2}, 'masses', rd.DEFAULTDATA)
-    {'U-238': 5.337817684684321e+22, 'Co-57': 7.61542657764305e+22}
+    >>> rd.inventory._input_to_number({'Ni-56': .3, 'Co-56': .7}, 'mBq', rd.DEFAULTDATA)
+    {'Ni-56': 227.17253191853973, 'Co-56': 6738.641562715049}
+    >>> rd.inventory._input_to_number({'U-238': 21.1, 'Co-57': 7.2}, 'μg', rd.DEFAULTDATA)
+    {'U-238': 5.337817684684321e+16, 'Co-57': 7.61542657764305e+16}
 
     """
 
-    number_dict = {}
-
     if units in activity_units:
-        for nuc, act in input_inv_dict.items():
-            number = _activity_to_number(
-                nuc, activity_unit_conv(act, units, "Bq"), data
-            )
-            number_dict[nuc] = number
-        converted_dict = number_dict.copy()
+        number_dict = {
+            nuc: _activity_to_number(nuc, activity_unit_conv(act, units, "Bq"), data)
+            for nuc, act in input_inv_dict.items()
+        }
     elif units in moles_units:
-        for nuc, moles in input_inv_dict.items():
-            number = _moles_to_number(moles_unit_conv(moles, units, "mol"))
-            number_dict[nuc] = number
-        converted_dict = number_dict.copy()
+        number_dict = {
[…]
