+++ tests/test_inventory.py
@@ -108,12 +108,36 @@ class Test(unittest.TestCase):
             {"Tc-99": 5.8, "Tc-99m": 2.3, "U-235": 1.2},
         )
 
+    def test__input_to_number(self):
+        """
+        Test the units conversion of inventory dictionaries from mass, moles, or activity into
+        number of atoms.
+        """
+
+        inv_dict = {"Ni-56": 0.3, "Co-56": 0.7}
+        self.assertEqual(
+            _input_to_number(inv_dict, "mBq", DEFAULTDATA),
+            {"Ni-56": 227.17253191853973, "Co-56": 6738.641562715049},
+        )
+        inv_dict = {"He-3": 1.0}
+        self.assertEqual(
+            _input_to_number(inv_dict, "mol", DEFAULTDATA),
+            {"He-3": 6.02214076e23},
+        )
+        inv_dict = {"U-238": 21.1, "Co-57": 7.2}
+        self.assertEqual(
+            _input_to_number(inv_dict, "μg", DEFAULTDATA),
+            {"U-238": 5.337817684684321e16, "Co-57": 7.61542657764305e16},
+        )
+
+        with self.assertRaises(ValueError):
+            _input_to_number((inv_dict), "xyz", DEFAULTDATA)
+
     def test__check_dictionary(self):
         """
         Test the checking of inventory dictionaries.
         """
 
-        nuclides = ["H-3", "C-14"]
         H3 = Radionuclide("H3")
         C14 = Radionuclide("C14")
 
@@ -158,6 +182,10 @@ class Test(unittest.TestCase):
             _check_dictionary({H3: 1.0, "C-14": 2.0}, units="num", data=DEFAULTDATA),
             {"C-14": 2.0, "H-3": 1.0},
         )
+        self.assertEqual(
+            _check_dictionary({"He-3": 1.0}, units="mol", data=DEFAULTDATA),
+            {"He-3": 6.02214076e23},
+        )
 
         # Catch incorrect arguments
         with self.assertRaises(ValueError):
@@ -167,14 +195,12 @@ class Test(unittest.TestCase):
 
     def test__sort_list_according_to_dataset(self):
         """
-        Test the sorting of list of radionuclides according to their position in the decay dataset.
+        Test the sorting of list of nuclides according to their position in the decay dataset.
         """
 
-        radionuclide_list = ["Tc-99", "Tc-99m"]
+        nuclide_list = ["Tc-99", "Tc-99m"]
         self.assertEqual(
-            _sort_list_according_to_dataset(
-                radionuclide_list, DEFAULTDATA.nuclide_dict
-            ),
+            _sort_list_according_to_dataset(nuclide_list, DEFAULTDATA.nuclide_dict),
             ["Tc-99m", "Tc-99"],
         )
 
@@ -197,6 +223,18 @@ class Test(unittest.TestCase):
         inv = Inventory({"Tc-99m": 2.3, I123: 5.8}, "num")
         self.assertEqual(inv.contents, {"Tc-99m": 2.3, "I-123": 5.8})
 
+        inv = Inventory({"H-3": 1.0})
+        self.assertEqual(inv.contents, {"H-3": 560892895.7794082})
+
+        inv = Inventory({"H-3": 1.0}, "mBq")
+        self.assertEqual(inv.contents, {"H-3": 560892.8957794083})
+
+        inv = Inventory({"He-3": 1.0}, "mol")
+        self.assertEqual(inv.contents, {"He-3": 6.02214076e23})
+
+        inv = Inventory({"He-3": 1.0}, "kg")
+        self.assertEqual(inv.contents, {"He-3": 1.9967116089131648e26})
+
[…]
