+++ docs/source/tutorial.rst
@@ -17,7 +17,7 @@ Create an inventory of radionuclides and associated activities as follows:
 
 .. code-block:: python3
 
-    >>> inv_t0 = rd.Inventory({'U-238': 99.274, 'U-235': 0.720, 'U-234': 0.005})
+    >>> inv_t0 = rd.Inventory({'U-238': 99.274, 'U-235': 0.720, 'U-234': 0.005}, 'mol')
 
 This is an inventory of natural uranium.
 
@@ -29,39 +29,41 @@ inventory:
 
 .. code-block:: python3
 
-    >>> inv_t0.activities()
-    {'U-234': 0.005, 'U-235': 0.72, 'U-238': 99.274}
+    >>> inv_t0.activities('MBq')
+    {'U-234': 269.4016050086039, 'U-235': 13.528246506057048, 'U-238': 293.90300567125}
     >>> inv_t0.radionuclides
     ['U-234', 'U-235', 'U-238']
-    >>> inv_t0.numbers():
-    {'U-234': 55884417984.51489, 'U-235': 2.3076736283495652e+16, 'U-238': 2.0193793783074922e+19}
-    >>> inv_t0.masses():
-    {'U-234': 2.1718592794623984e-11, 'U-235': 9.006841520909447e-06, 'U-238': 0.00798245788808877}
-    >>> inv_t0.mass_abundances():
-    {'U-234': 2.7177236552303014e-09, 'U-235': 0.0011270576547825488, 'U-238': 0.9988729396274937}
-    >>> inv_t0.moles():
-    {'U-234': 9.279825930955971e-14, 'U-235': 3.831982214161273e-08, 'U-238': 3.3532583491248255e-05}
+    >>> inv_t0.numbers()
+    {'U-234': 3.01107038e+21, 'U-235': 4.3359413471999995e+23, 'U-238': 5.9784200180824e+25}
+    >>> inv_t0.masses('g')
+    {'U-234': 1.17020475148, 'U-235': 169.23162824424, 'U-238': 23632.253822284463}
+    >>> inv_t0.mass_fractions()
+    {'U-234': 4.916278117985593e-05, 'U-235': 0.007109779290811992, 'U-238': 0.992841057928008}
+    >>> inv_t0.moles()
+    {'U-234': 0.005, 'U-235': 0.72, 'U-238': 99.274}
+    >>> inv_t0.mole_fractions()
+    {'U-234': 5.000050000500006e-05, 'U-235': 0.0072000720007200075, 'U-238': 0.992749927499275}
     >>> inv_t0
-    Inventory: {'U-234': 55884417984.51489, 'U-235': 2.3076736283495652e+16, 'U-238': 2.0193793783074922e+19}, decay dataset: icrp107
+    Inventory activities (Bq): {'U-234': 269401605.0086039, 'U-235': 13528246.506057048, 'U-238': 293903005.67125}, decay dataset: icrp107
 
-The input of ``Inventory`` is by default activity. The user can easily change
-the input using the ``input_type`` keyword argument during initialization: 
+By default the input dictionary to ``Inventory()`` is assumed to contain
+activities in Bq. The user can easily specify different units:
 
 .. code-block:: python3
 
     # initialize an inventory using activities:
-    >>> inv1 = rd.Inventory({'U-238': 99.274, 'U-235': 0.720, 'U-234': 0.005})
-    >>> inv1.activities()
-    {'U-234': 0.005, 'U-235': 0.72, 'U-238': 99.274}
+    >>> inv1 = rd.Inventory({'C-14': 5.0, 'H-3': 2.0})
[…]

[…]
