### 2021-08-04 c034df0
Move utils to Nuclide class, includes Z, A parsing
