### 2021-08-16 #3318 @munkm
@kellyrowland i've been reviewing this package and I have a question about the LICENSE file. The main license is using the MIT language, but then there are supplements to the license for the nuclear data ICRP-07 that amend it. The additions to the license should allow the package to redistribute the data for educational, not for profit, or research purposes but I am unsure if this is an issue because it does seem to fall outside of the other license texts? What do you think?
