### 2021-08-16 #3318 @kellyrowland
Good catch, @munkm .

I think that the license text should be separated out into two files: one file with the MIT license text for this software package, and a second, separate license file which includes the wording for the data libraries and their respective (more restrictive) license texts. That should hopefully help differentiate the licenses.

Does that make sense? Does that work for you @alexmalins ?
