### 2021-08-16 alexmalins/radioactivedecay#37 @munkm
JOSS Paper: Minor edits to the paper
Hi,

I'm reviewing this tool for JOSS in https://github.com/openjournals/joss-reviews/issues/3318 and I have a few minor suggestions that I think would improve the paper:

- [x] add citations to PyNE and Radiological toolbox on line 25. You have them later but this is their first appearance in the manuscript
- [x] on l86 you mention fig1a is based on NetworkX and l91 is based on Matplotlib, but I think it would be clearer if you said it is made using those tools.
- [x] I do not believe PyNE can create decay diagrams like fig 1a, so you could add that feature to your statement of need section too. They're very nice.
- [x] It might be worth explicitly stating that all of this is based on atomic fractions and not mass fractions of the nuclear materials, unless there's an option for that that I missed in the documentation. This isn't necessary for the JOSS publication though
