### 2021-08-16 alexmalins/radioactivedecay#38 @munkm
JOSS Documentation
The documentation for this package is nicely written and guides potential users through the software and API well. Based on the reviewer checklist in https://github.com/openjournals/joss-reviews/issues/3318, I think a few minor enhancements can be made.

- [x] **statement of need** You mention what the package does, but you don't explicitly state who the target audience for the software is in the documentation
- [x] **community guidelines** Do you have a preferred way for users to seek out support or ask questions? There are instructions for feature requests and bugfixes, but what about questions outside of these?
- [x] **testing** The package does have a tests directory, but I don't see on https://alexmalins.com/radioactivedecay/ that there are steps described on how to run the tests to check one's installation. Did I miss it?
