### 2021-08-17 #3318 @alexmalins
@munkm & @kellyrowland

Thanks for your review and the suggestions.

Honestly, when I created the license file I wasn't sure of the best way to do it. Splitting the license file into separate files for the code (MIT) and the default data it uses (ICRP-07) works for me. Moreover, this makes sense as the code is programmed flexibly to use other datasets, and there is the potential that the default will switch to something more recent in the future.

Thanks for the issues you created in the repository, @munkm. I'll work on these and splitting the license files in the coming days.
