### 2021-08-20 10a7ddf
Very large refactor
