+++ radioactivedecay/inventory.py
@@ -1,10 +1,10 @@
 """
-The inventory module defines the ``Inventory`` class. Each ``Inventory`` instance contains one or
-more nuclide each with an associated quantity (number of atoms). The decay of the nuclide(s) in an
+The inventory module defines the ``Inventory`` class. An ``Inventory`` instance contains one or
+more nuclides each with an associated amount (number of atoms). The decay of the nuclide(s) in an
 ``Inventory`` can be calculated by using the ``decay()`` method (normal double-precision
-floating-point operations). Use the ``decay_high_precision()`` method to perform a SymPy high
-numerical precision calculation. A ``DecayData`` dataset is associated with each ``Inventory``
-instance (default is rd.DEFAULTDATA).
+floating-point operations). The corresponding ``InventoryHP`` class performs decay calculations
+with SymPy high numerical precision operations. A ``DecayData`` dataset is associated with
+``Inventory`` and ``InventoryHP`` instances (default is rd.DEFAULTDATA).
 
 The code examples shown in the docstrings assume the ``radioactivedecay`` package has been imported
 as:
@@ -17,420 +17,35 @@ as:
 """
 
 from functools import singledispatch, update_wrapper
-from typing import Callable, Dict, List, Tuple, Union
+from typing import Callable, Dict, List, Optional, Tuple, Union
 import matplotlib
-from sympy import exp, nsimplify
-from scipy.constants import Avogadro
-from radioactivedecay.decaydata import DecayData, DEFAULTDATA, np
+import numpy as np
+from scipy import sparse
+from sympy import exp, Matrix, nsimplify
+from sympy.core.expr import Expr
+from radioactivedecay.converters import (
+    QuantityConverter,
+    QuantityConverterSympy,
+    UnitConverter,
+    UnitConverterSympy,
+)
+from radioactivedecay.decaydata import (
+    DecayData,
+    DecayMatrices,
+    DecayMatricesSympy,
+    DEFAULTDATA,
+)
 from radioactivedecay.plots import _decay_graph
 from radioactivedecay.radionuclide import Radionuclide
 from radioactivedecay.utils import (
-    activity_units,
-    activity_unit_conv,
-    mass_units,
-    mass_unit_conv,
-    moles_units,
-    moles_unit_conv,
     parse_nuclide,
-    time_unit_conv,
-    time_unit_conv_sympy,
+    add_dictionaries,
+    sort_dictionary_alphabetically,
+    sort_list_according_to_dataset,
 )
 
 
 # pylint: disable=too-many-arguments, too-many-locals
-
-
-def _add_dictionaries(
-    dict_a: Dict[str, float], dict_b: Dict[str, float]
-) -> Dict[str, float]:
-    """
-    Adds together two dictionaries of nuclides and associated quantities.
-
-    Parameters
-    ----------
-    dict_a : dict
-        First dictionary containing nuclide strings as keys and quantitites as values.
-    dict_b : dict
-        Second dictionary containing nuclide strings as keys and quantitites as values.
-
-    Returns
-    -------
-    dict
-        Combined dictionary containing the nuclides in both dict_a and dict_b, where the
-        quantities have been added together when a nuclide is present in both input
-        dictionaries.
-
-    Examples
-    --------
-    >>> dict_a = {'Pm-141': 1.0, 'Rb-78': 2.0}
-    >>> dict_b = {'Pm-141': 3.0, 'Rb-90': 4.0}
-    >>> rd.inventory._add_dictionaries(dict_a, dict_b)
-    {'Pm-141': 4.0, 'Rb-78': 2.0, 'Rb-90': 4.0}
-
-    """
-
-    new_dict = dict_a.copy()
[…]
