### 2021-08-21 16d4559
Refactor UnitConverter using abc, add test for InventoryHP.numbers()
