### 2021-08-23 3f37019
Add citations l25, 'based on' -> 'using' for NetworkX and Matplotlib
