+++ docs/source/index.rst
@@ -12,12 +12,18 @@ Welcome to radioactivedecay's documentation!
 ``radioactivedecay`` is a Python package for radioactive decay calculations.
 It supports decay chains of radionuclides, metastable states and branching
 decays. By default it uses the decay data from ICRP Publication 107, which
-contains 1252 radionuclides of 97 elements.
+contains 1252 radionuclides of 97 elements, and atomic mass data from the
+Atomic Mass Data Center.
 
-It solves the radioactive decay differential equations analytically using NumPy
-and SciPy linear algebra routines. There is also a high numerical precision
-mode using SymPy routines which gives more accurate results for decay chains
-with orders of magnitude differences between radionuclide half-lives.
+The code solves the radioactive decay differential equations analytically using
+NumPy and SciPy linear algebra routines. There is also a high numerical
+precision calculation mode employing SymPy routines. This gives more accurate
+results for decay chains containing radionuclides with orders of magnitude
+differences between the half-lives.
+
+This is free-to-use open source software. It was created for engineers,
+technicians and researchers who work with and study radioactivity, and for
+educational use.
 
 .. toctree::
    :maxdepth: 2
