### 2021-08-23 alexmalins/radioactivedecay#37 @alexmalins
Hi @munkm,

Thanks for these suggestions and ideas. Over on the [joss_paper](https://github.com/alexmalins/radioactivedecay/tree/joss_paper) branch I committed changes in line with the first two suggestions.

Item 3: Actually when I was developing this function I stumbled across this [script](https://github.com/pyne/pyne/blob/9c1d41237dc26597cd2463d5968d2ab4e2cfce0f/pyne/gui/decaychain.py) in the PyNE repository which appears to create decay diagrams using Graphviz. I've not tried it - it doesn't appear to be documented anywhere and I'm not sure if it is supported. However given it exists and because Radiological Toolbox can plot similar diagrams, I didn't feel this was a sufficiently distinguishing feature to add to the statement of need in the radioactivedecay JoSS paper.

Item 4: All of this is based on activity / activity fractions in the version that was entered for review (v0.3.2). As it happens, a student (Thom Lemoine - @lemointm) who has been working on [TARDIS](https://github.com/tardis-sn/tardis) & radioactivedecay over summer has coded new functionality so radioactivedecay supports mass, mass fractions, moles, number of atoms & number fractions. This is all now working over on the [devel](https://github.com/alexmalins/radioactivedecay/tree/devel) branch, and we're close to cutting a new release. Instead of stating this limitation of the old version, I was wondering if there was a way to add Thom as a co-author to the paper so a) he can get credit for this good work, and b) the JoSS paper is up to date with the latest functionality?
