### 2021-08-23 alexmalins/radioactivedecay#38 @alexmalins
Hi @munkm,

Thanks for your review and these suggestions. I incorporated them all. The changes are visible within the [doc/source](https://github.com/alexmalins/radioactivedecay/tree/devel/docs/source) folder of the devel branch (apologies the commit history is not so clear as some of the changes were incorporated as part of a large refactor). Once we release the next version, the changes will go live on https://alexmalins.com/radioactivedecay/.
