### 2021-08-23 #3318 @alexmalins
Progress update: I've committed changes incorporating your suggestions @munkm and @kellyrowland . These are currently visible on the [devel](https://github.com/alexmalins/radioactivedecay/tree/devel) and [joss_paper](https://github.com/alexmalins/radioactivedecay/tree/joss_paper) branches of the repository.

There are two outstanding comments from the review in this [issue](https://github.com/alexmalins/radioactivedecay/issues/37). The main one is item 4, where @munkm asked for the inclusion of a statement about code only dealing with activities. Coincidentally, @lemointm worked on improvements this summer so the code now supports masses, moles and number of atoms of radioactive materials. These will be incorporated in the next release, which is basically ready to go live. I was wondering if it would be allowed to update the paper along these lines and add Thom Lemoine as a co-author?
