### 2021-08-24 a5fc901
Add activity_fractions() method for inventory classes, refactor test imports to fix mypy errors
