### 2021-08-24 182220f
Add conda and code style badges
