### 2021-08-25 64f7074
Add option to plot activity fractions on with inventory plot() methods
