### 2021-08-25 f992d8e
Absorb Radionuclide into Nuclide class, add masses
