+++ radioactivedecay/nuclide.py
@@ -1,10 +1,13 @@
 """
 The nuclide module defines the ``Nuclide`` class. Each ``Nuclide``
 instance represents one nuclide or element, and has a variety of ways
-for contruction based off of a nuclide or element string, or atomic
-number and mass. The class properties provide an easy way of
-extracting atomic number and mass information, as well as a clean name
-string.
+for contruction based off of a nuclide or element string, atomic
+number and mass, or `zzzaaassss` canonical id. The class properties
+provide an easy way of extracting atomic number and mass information,
+as well as a clean name string. Furthermore, additional methods provide
+an access point for mass data of nuclides, and the decay data of
+radionuclides, if present in a specified dataset. The default decay
+dataset used if none is supplied to the constructor is rd.DEFAULTDATA.
 
 The code examples shown in the docstrings assume the
 ``radioactivedecay`` package has been imported as:
@@ -16,8 +19,17 @@ The code examples shown in the docstrings assume the
 
 """
 
-import os, csv
-from typing import Any
+from collections import deque
+from typing import Any, Dict, List, Tuple, Union
+import matplotlib
+import networkx as nx
+from radioactivedecay.decaydata import DecayData, DEFAULTDATA
+from radioactivedecay.plots import (
+    _parse_nuclide_label,
+    _parse_decay_mode_label,
+    _check_fig_ax,
+)
+from radioactivedecay.utils import parse_radionuclide
 from radioactivedecay.utils import parse_nuclide
 
 
@@ -60,28 +72,62 @@ class Nuclide:
 
     """
     
-    def __init__(self, input: Any, A: int = 0) -> None:
-        loc = os.path.realpath(
-            os.path.join(os.getcwd(), os.path.dirname(__file__))
-        )
-        sym_dict_path = os.path.join(loc, "element_symbols.csv")
-
-        Z_dict, sym_dict = {}, {}
-        with open(sym_dict_path, 'r') as f:
-            csv_list = csv.DictReader(f)
-            for row in csv_list:
-                Z_dict[int(row["Z"])] = row["Symbol"]
-                sym_dict[row["Symbol"]] = int(row["Z"])
+    def __init__(
+        self, input: Any,
+        A: int = 0,
+        meta_state: str = "",
+        data: DecayData = DEFAULTDATA
+        ) -> None:
+        
+        Z_dict = {
+            1: "H", 2: "He", 3: "Li", 4: "Be", 5: "B", 6: "C", 7: "N", 8: "O",
+            9: "F", 10: "Ne", 11: "Na", 12: "Mg", 13: "Al", 14: "Si", 15: "P",
+            16: "S", 17: "Cl", 18: "Ar", 19: "K", 20: "Ca", 21: "Sc", 22: "Ti",
+            23: "V", 24: "Cr", 25: "Mn", 26: "Fe", 27: "Co", 28: "Ni", 29: "Cu",
+            30: "Zn", 31: "Ga", 32: "Ge", 33: "As", 34: "Se", 35: "Br", 36: "Kr",
+            37: "Rb", 38: "Sr", 39: "Y", 40: "Zr", 41: "Nb", 42: "Mo", 43: "Tc",
+            44: "Ru", 45: "Rh", 46: "Pd", 47: "Ag", 48: "Cd", 49: "In", 50: "Sn",
+            51: "Sb", 52: "Te", 53: "I", 54: "Xe", 55: "Cs", 56: "Ba", 57: "La",
[…]
