### 2021-08-25 aa148b0
Small typo fixes for decaydata.py and utils.py
