### 2021-08-25 72e39d4
add inventory cumulative_decays() method, fix InventoryHP.decay() docstring, readme header reformat
