+++ tests/test_inventory.py
@@ -440,6 +440,74 @@ class TestInventory(unittest.TestCase):
             },
         )
 
+    def test_cumulative_decays(self) -> None:
+        """
+        Test Inventory.cumulative_decays() calculations.
+        """
+
+        inv = Inventory({"H-3": 10.0}, "num")
+        self.assertEqual(inv.cumulative_decays(12.32, "y"), {"H-3": 5.0})
+        self.assertEqual(inv.cumulative_decays(1e6, "y"), {"H-3": 10.0})
+
+        inv = Inventory({"Sr-90": 10.0}, "num")
+        self.assertEqual(
+            inv.cumulative_decays(1e6, "y"), {"Sr-90": 10.0, "Y-90": 10.000000000000002}
+        )
+
+        inv = Inventory({"Tc-99m": 2.3, "I-123": 5.8}, "num")
+        self.assertEqual(
+            inv.cumulative_decays(20.0, "h"),
+            {
+                "I-123": 3.759540755465226,
+                "Te-123m": 4.72801274418656e-07,
+                "Te-123": -9.485843969524933e-18,
+                "Tc-99m": 2.0704925198993647,
+                "Tc-99": 1.0500063924036559e-08,
+            },
+        )
+
+        inv = Inventory({"U-238": 99.274, "U-235": 0.720, "U-234": 0.005}, "num")
+        self.assertEqual(
+            inv.cumulative_decays(1e9, "y"),
+            {
+                "Ac-227": 0.45099937182594435,
+                "Th-227": 0.44477558047547366,
+                "Ra-226": 14.264656807713404,
+                "Fr-223": 0.006223791331197811,
+                "Ra-223": 0.4509993717947137,
+                "Rn-222": 14.264656807514221,
+                "Rn-219": 0.4509993717947137,
+                "At-219": 3.734274798718682e-07,
+                "Po-218": 14.264656807514106,
+                "At-218": 0.002852931361502821,
+                "Rn-218": 2.852931361502822e-06,
+                "Bi-215": 3.622246554757077e-07,
+                "Po-215": 0.4509997340193692,
+                "Pb-214": 14.261803876151637,
+                "Bi-214": 14.264653954581055,
+                "Po-214": 14.261661230181954,
+                "Pb-211": 0.450999734019343,
+                "Bi-211": 0.4509997340193414,
+                "Po-211": 0.0012447592658933822,
+                "Tl-210": 0.0029955773304620116,
+                "Pb-210": 14.264656385104534,
+                "Bi-210": 14.264656384843379,
+                "Po-210": 14.26465637763465,
+                "U-238": 14.265792678151335,
+                "Tl-207": 0.4497549747534446,
+                "Hg-206": 2.710284713169805e-07,
+                "Tl-206": 1.910037489931004e-05,
+                "U-235": 0.4510118978455057,
+                "Th-234": 14.26579267689593,
+                "Pa-234m": 14.265792676895886,
+                "Pa-234": 0.022825268283010153,
+                "U-234": 14.266121535073315,
+                "Th-231": 0.45101189784439344,
+                "Pa-231": 0.45099938014509444,
+                "Th-230": 14.264687251524753,
+            },
+        )
+
     def test_half_lives(self) -> None:
         """
         Test method to fetch half-lives of nuclides in the Inventory.
@@ -692,6 +760,72 @@ class TestInventoryHP(unittest.TestCase):
                 "U-234": 85.01287846492669,
                 "U-235": 0.26898810215449415,
                 "U-238": 85.00820732184867,
+            },
+        )
+
+    def test_cumulative_decays(self) -> None:
+        """
[…]
