### 2021-08-28 0ebae8e
Element parsing to utils, add Nuclide.atomic_mass
