+++ radioactivedecay/nuclide.py
@@ -1,13 +1,12 @@
 """
 The nuclide module defines the ``Nuclide`` class. Each ``Nuclide``
-instance represents one nuclide or element, and has a variety of ways
-for contruction based off of a nuclide or element string, atomic
-number and mass, or `zzzaaassss` canonical id. The class properties
-provide an easy way of extracting atomic number and mass information,
-as well as a clean name string. Furthermore, additional methods provide
-an access point for mass data of nuclides, and the decay data of
-radionuclides, if present in a specified dataset. The default decay
-dataset used if none is supplied to the constructor is rd.DEFAULTDATA.
+instance represents one nuclide, and can be contructed from a nuclide
+string or `zzzaaassss` canonical id. The class properties provide an
+easy way of extracting atomic number and mass information, as well as a
+clean name string. Furthermore, additional methods provide an access
+point for mass data of nuclides, and the decay data of radionuclides,
+if present in a specified dataset. The default decay dataset used if
+none is supplied to the constructor is rd.DEFAULTDATA.
 
 The code examples shown in the docstrings assume the
 ``radioactivedecay`` package has been imported as:
@@ -29,8 +28,12 @@ from radioactivedecay.plots import (
     _parse_decay_mode_label,
     _check_fig_ax,
 )
-from radioactivedecay.utils import parse_radionuclide
-from radioactivedecay.utils import parse_nuclide
+from radioactivedecay.utils import (
+    parse_nuclide,
+    elem_to_Z,
+    build_id,
+    build_nuclide_string
+)
 
 
 class Nuclide:
@@ -42,20 +45,19 @@ class Nuclide:
     ----------
     input : Any
         Input value for instantiation. Can be nuclide string in name
-        format (with or without hyphen), or atomic number (Z).
-    A : int, optional
-        For atomic number as input, atomic mass can be optionally
-        added for full nuclide string construction. 
+        format (with or without hyphen), or canonical id (zzzaaassss).
 
-    Properties
+    Attributes
     ----------
     name : str
-        Nuclide string if A data is given, otherwise the element
-        symbol is returned.
+        Nuclide name string.
     Z : int
         Atomic number.
     A : int
-        Atomic mass, returns 0 for raw elements.
+        Atomic mass number.
+    id : int
+        Canonical nuclide id, in zzzaaassss form. Ground state is 0000,
+        first excited state ("m") is 0001, second ("n") is 0002, etc.
 
     Examples
     --------
@@ -63,192 +65,66 @@ class Nuclide:
     Nuclide: K-40
     >>> rd.Nuclide('K40')
     Nuclide: K-40
-    >>> rd.Nuclide('K')
-    Element: K
-    >>> rd.Nuclide(19, 40)
+    >>> rd.Nuclide(190400000)
     Nuclide: K-40
-    >>> rd.Nuclide(19)
-    Element: K
+    >>> rd.Nuclide(280560001)
+    Nuclide: Ni-56m
 
     """
     
     def __init__(
-        self, input: Any,
-        A: int = 0,
-        meta_state: str = "",
+        self,
+        input: Any,
         data: DecayData = DEFAULTDATA
-        ) -> None:
-        
-        Z_dict = {
-            1: "H", 2: "He", 3: "Li", 4: "Be", 5: "B", 6: "C", 7: "N", 8: "O",
-            9: "F", 10: "Ne", 11: "Na", 12: "Mg", 13: "Al", 14: "Si", 15: "P",
[…]
