+++ radioactivedecay/nuclide.py
@@ -19,14 +19,15 @@ The code examples shown in the docstrings assume the
 """
 
 from collections import deque
-from typing import Any, Dict, List, Tuple, Union
+from typing import Any, Dict, List, Tuple, Union, Optional
 import matplotlib
 import networkx as nx
+import numpy as np
 from radioactivedecay.decaydata import DecayData, DEFAULTDATA
 from radioactivedecay.plots import (
     _parse_nuclide_label,
     _parse_decay_mode_label,
-    _check_fig_ax,
+    _check_fig_axes,
 )
 from radioactivedecay.utils import (
     parse_nuclide,
@@ -46,10 +47,12 @@ class Nuclide:
     input : Any
         Input value for instantiation. Can be nuclide string in name
         format (with or without hyphen), or canonical id (zzzaaassss).
+    decay_data : DecayData, optional
+        Decay dataset (default is the ICRP-107 dataset).
 
     Attributes
     ----------
-    name : str
+    nuclide : str
         Nuclide name string.
     Z : int
         Atomic number.
@@ -58,6 +61,12 @@ class Nuclide:
     id : int
         Canonical nuclide id, in zzzaaassss form. Ground state is 0000,
         first excited state ("m") is 0001, second ("n") is 0002, etc.
+
+    Raises
+    ------
+    ValueError
+        If the input nuclide string is invalid or the nuclide is not
+        contained in the decay dataset.
 
     Examples
     --------
@@ -75,10 +84,16 @@ class Nuclide:
     def __init__(
         self,
         input: Any,
-        data: DecayData = DEFAULTDATA
+        decay_data: DecayData = DEFAULTDATA
     ) -> None:
+        self.decay_data = decay_data
         self.parse_name(input)
-        self.data = data
+        index = decay_data.nuclide_dict[self.nuclide]
+        mass = decay_data.scipy_data.atomic_masses[index]
+        self.atomic_mass = mass
+        self.prog_bf_mode: Dict[
+            str, List[Union[float, str]]
+        ] = decay_data.prog_bfs_modes[decay_data.nuclide_dict[self.nuclide]]
 
     def parse_name(self, input: Any) -> None:
         if isinstance(input, int):
@@ -92,9 +107,17 @@ class Nuclide:
                 state = "n"
             Z = int(id_zzzaaa / 1000)
             A = id_zzzaaa - (Z * 1000)    
-            name = build_nuclide_string(Z, A, state)
+            name = parse_nuclide(
+                build_nuclide_string(Z, A, state),
+                self.decay_data.nuclides,
+                self.decay_data.dataset_name
+            )
         elif isinstance(input, str):
-            name = parse_nuclide(input)
+            name = parse_nuclide(
+                input,
+                self.decay_data.nuclides,
+                self.decay_data.dataset_name
+            )
             Z = elem_to_Z(name.split("-")[0])
             A = int(name.split("-")[1].strip("mn"))
             state = name.split("-")[1].strip("0123456789")
@@ -104,53 +127,18 @@ class Nuclide:
                 "Invalid input type, expected int or str"
             )
         
-        self.name = name
+        self.nuclide = name
         self.Z = Z
         self.A = A
         self.state = state
     
     def __repr__(self) -> str:
-        if self.A == 0:
-            rep =  "Element: " + self.name
-        else:
-            rep = ("Nuclide: "
-                   + self.name
-                   + ", decay dataset: "
-                   + self.data.dataset)
+        rep = ("Nuclide: "
+            + self.nuclide
+            + ", decay dataset: "
+            + self.decay_data.dataset_name)
             
         return rep
-    
-    @property
-    def atomic_mass(self) -> int:
-        """
-        Returns the atomic weight of the nuclide, if contained in the
-        dataset.
-        
-        Raises
-        ------
[…]
