+++ radioactivedecay/inventory.py
@@ -36,7 +36,7 @@ from radioactivedecay.decaydata import (
     DEFAULTDATA,
 )
 from radioactivedecay.plots import _decay_graph
-from radioactivedecay.radionuclide import Radionuclide
+from radioactivedecay.nuclide import Nuclide
 from radioactivedecay.utils import (
     parse_nuclide,
     add_dictionaries,
@@ -69,7 +69,7 @@ class Inventory:
     Parameters
     ----------
     contents : dict
-        Dictionary containing nuclide strings or Radionuclide instances as keys and quantities
+        Dictionary containing nuclide strings or Nuclide instances as keys and quantities
         (activities, number of atoms, masses or moles) as values.
     units : str, optional
         Units of the contents dictionary values. Specify either 'num' for number of atoms, or an
@@ -99,7 +99,7 @@ class Inventory:
     --------
     >>> rd.Inventory({'Tc-99m': 2.3, 'I-123': 5.8}, 'Bq')
     Inventory activities (Bq): {'I-123': 2.3, 'Tc-99m': 5.8}, decay dataset: icrp107_ame2020_nubase2020
-    >>> H3 = rd.Radionuclide('H-3')
+    >>> H3 = rd.Nuclide('H-3')
     >>> rd.Inventory({H3: 3.0}, 'g')
     Inventory activities (Bq): {'H-3': 1067957043281807.0}, decay dataset: icrp107_ame2020_nubase2020
     >>> rd.Inventory({'U-238': 21.1, 'Co-57': 7.2}, 'Ci')
@@ -109,7 +109,7 @@ class Inventory:
 
     def __init__(
         self,
-        contents: Dict[Union[str, Radionuclide], float],
+        contents: Dict[Union[str, Nuclide], float],
         units: str = "Bq",
         check: bool = True,
         decay_data: DecayData = DEFAULTDATA,
@@ -134,19 +134,19 @@ class Inventory:
 
     @staticmethod
     def _parse_nuclides(
-        contents: Dict[Union[str, Radionuclide], Union[float, Expr]],
+        contents: Dict[Union[str, Nuclide], Union[float, Expr]],
         nuclides: List[str],
         dataset_name: str,
     ) -> Dict[str, Union[float, Expr]]:
         """
-        Checks that nuclide keys in the contents dictionary. Converts Radionuclide instances into
+        Checks that nuclide keys in the contents dictionary. Converts Nuclide instances into
         nuclide name strings. Converts nuclide name strings into Ab-XY format.
         """
 
         return {
             (
                 parse_nuclide(nuc.nuclide, nuclides, dataset_name)
-                if isinstance(nuc, Radionuclide)
+                if isinstance(nuc, Nuclide)
                 else parse_nuclide(nuc, nuclides, dataset_name)
             ): inp
             for nuc, inp in contents.items()
@@ -191,7 +191,7 @@ class Inventory:
         Parameters
         ----------
         contents : dict
-            Dictionary containing nuclide strings or Radionuclide objects as keys, and masses,
+            Dictionary containing nuclide strings or Nuclide objects as keys, and masses,
             moles, or activities as values.
         units : str
             Units of the values in the input dictionary.
@@ -412,7 +412,7 @@ class Inventory:
 
     def add(
         self,
-        add_contents: Dict[Union[str, Radionuclide], float],
+        add_contents: Dict[Union[str, Nuclide], float],
         units: str = "Bq",
     ) -> None:
         """
@@ -422,7 +422,7 @@ class Inventory:
         Parameters
         ----------
         add_contents : dict
-            Dictionary containing nuclide strings or Radionuclide objects as keys
+            Dictionary containing nuclide strings or Nuclide objects as keys
             and the amount of each nuclide (with specified units) as values
[…]
