+++ radioactivedecay/__init__.py
@@ -19,6 +19,5 @@ educational use.
 __version__ = "0.4.0"
 
 from radioactivedecay.decaydata import DecayData, DEFAULTDATA
-from radioactivedecay.radionuclide import Radionuclide
 from radioactivedecay.inventory import Inventory, InventoryHP
 from radioactivedecay.nuclide import Nuclide
