+++ tests/test_nuclide.py
@@ -1,128 +1,128 @@
 """
-Unit tests for radionuclide.py functions, classes and methods.
+Unit tests for nuclide.py functions, classes and methods.
 """
 
 import unittest
 from radioactivedecay.decaydata import DecayData
-from radioactivedecay.radionuclide import Radionuclide
+from radioactivedecay.nuclide import Nuclide
 
 
-class TestRadionuclide(unittest.TestCase):
+class TestNuclide(unittest.TestCase):
     """
-    Unit tests for the radionuclide.py Radionuclide class.
+    Unit tests for the nuclide.py Nuclide class.
     """
 
-    def test_radionuclide_instantiation(self) -> None:
+    def test_nuclide_instantiation(self) -> None:
         """
-        Test instantiation of Radionuclide objects.
+        Test instantiation of Nuclide objects.
         """
 
-        nuc = Radionuclide("Rn-222")
+        nuc = Nuclide("Rn-222")
         self.assertEqual(nuc.nuclide, "Rn-222")
         self.assertEqual(nuc.prog_bf_mode, {"Po-218": [1.0, "\u03b1"]})
 
     def test_radionuclide_half_life(self) -> None:
         """
-        Test Radionuclide half_life() method.
+        Test Nuclide half_life() method.
         """
 
-        nuc = Radionuclide("H-3")
+        nuc = Nuclide("H-3")
         self.assertEqual(nuc.half_life(), 388781329.30560005)
         self.assertEqual(nuc.half_life("y"), 12.32)
         self.assertEqual(nuc.half_life("readable"), "12.32 y")
 
     def test_radionuclide_progeny(self) -> None:
         """
-        Test Radionuclide half_life() method.
+        Test Nuclide half_life() method.
         """
 
-        nuc = Radionuclide("K-40")
+        nuc = Nuclide("K-40")
         self.assertEqual(nuc.progeny()[0], "Ca-40")
         self.assertEqual(nuc.progeny()[1], "Ar-40")
 
     def test_radionuclide_branching_fractions(self) -> None:
         """
-        Test Radionuclide branching_fractions() method.
+        Test Nuclide branching_fractions() method.
         """
 
-        nuc = Radionuclide("K-40")
+        nuc = Nuclide("K-40")
         self.assertEqual(nuc.branching_fractions()[0], 0.8914)
         self.assertEqual(nuc.branching_fractions()[1], 0.1086)
 
     def test_radionuclide_decay_modes(self) -> None:
         """
-        Test Radionuclide decay_modes() method.
+        Test Nuclide decay_modes() method.
         """
 
-        nuc = Radionuclide("K-40")
+        nuc = Nuclide("K-40")
         self.assertEqual(nuc.decay_modes()[0], "\u03b2-")
         self.assertEqual(nuc.decay_modes()[1], "\u03b2+ & EC")
 
     def test_radionuclide_plot(self) -> None:
         """
-        Test Radionuclide plot() method.
+        Test Nuclide plot() method.
 
         Only testing auto-generation of limits so far.
         """
 
-        nuc = Radionuclide("H-3")
+        nuc = Nuclide("H-3")
         _, axes = nuc.plot()
         self.assertEqual(axes.get_xlim(), (-0.3, 0.3))
         self.assertEqual(axes.get_ylim(), (-1.3, 0.3))
 
-        nuc = Radionuclide("Mo-99")
+        nuc = Nuclide("Mo-99")
         _, axes = nuc.plot()
         self.assertEqual(axes.get_xlim(), (-0.3, 1.3))
         self.assertEqual(axes.get_ylim(), (-2.3, 0.3))
 
[…]
