### 2021-09-02 7c5fabb
Delete radionuclide.py
Deleting as functionality has been moved to the `Nuclide` class
