+++ radioactivedecay/radionuclide.py
@@ -1,389 +1,0 @@
-"""
-The radionuclide module defines the ``Radionuclide`` class. Each ``Radionuclide`` instance
-represents one nuclide from the associated ``DecayData`` dataset. The methods provide an
-access point for the decay data. The default decay dataset used if none is supplied to the
-constructor is rd.DEFAULTDATA.
-
-The code examples shown in the docstrings assume the ``radioactivedecay`` package has been imported
-as:
-
-.. highlight:: python
-.. code-block:: python
-
-    >>> import radioactivedecay as rd
-
-"""
-
-from collections import deque
-from typing import Any, Dict, List, Optional, Tuple, Union
-import matplotlib
-import networkx as nx
-import numpy as np
-from radioactivedecay.decaydata import DecayData, DEFAULTDATA
-from radioactivedecay.plots import (
-    _parse_nuclide_label,
-    _parse_decay_mode_label,
-    _check_fig_axes,
-)
-from radioactivedecay.utils import parse_nuclide
-
-
-class Radionuclide:
-    """
-    ``Radionuclide`` instances represent one nuclide from the assoicated ``DecayData``
-    dataset.
-
-    Parameters
-    ----------
-    nuclide : str
-        Nuclide string.
-    decay_data : DecayData, optional
-        Decay dataset (default is the ICRP-107 dataset).
-
-    Attributes
-    ----------
-    nuclide : str
-        nuclide string.
-    prog_bf_mode : dict
-        Dictionary containing direct progeny as keys, and a list containing the branching fraction
-        and the decay mode for that progeny as values.
-    decay_data : DecayData
-        Decay dataset.
-
-    Examples
-    --------
-    >>> rd.Radionuclide('K-40')
-    Nuclide: K-40, decay dataset: icrp107
-
-    """
-
-    # pylint: disable=too-many-arguments
-
-    def __init__(self, nuclide: str, decay_data: DecayData = DEFAULTDATA) -> None:
-        self.nuclide: str = parse_nuclide(
-            nuclide, decay_data.nuclides, decay_data.dataset_name
-        )
-        self.prog_bf_mode: Dict[
-            str, List[Union[float, str]]
-        ] = decay_data.prog_bfs_modes[decay_data.nuclide_dict[self.nuclide]]
-        self.decay_data: DecayData = decay_data
-
-    def half_life(self, units: str = "s") -> Union[float, str]:
-        """
-        Returns the half-life of the nuclide as a float in your chosen units, or as
-        a human-readable string with appropriate units.
-
-        Parameters
-        ----------
-        units : str, optional
-            Units for half-life. Options are 'ps', 'ns', 'μs', 'us', 'ms', 's', 'm', 'h', 'd', 'y',
-            'ky', 'My', 'By', 'Gy', 'Ty', 'Py', and common spelling variations. Default is 's', i.e.
-            seconds. Use 'readable' to get a string of the half-life in human-readable units.
-
-        Returns
-        -------
-        float or str
-            Nuclide half-life.
-
-        Examples
-        --------
-        >>> K40 = rd.Radionuclide('K-40')
-        >>> K40.half_life('y')
-        1251000000.0
-        >>> K40.half_life('readable')
-        '1.251 By'
-
-        """
-
-        return self.decay_data.half_life(self.nuclide, units)
-
-    def progeny(self) -> List[str]:
-        """
-        Returns the direct progeny of the nuclide.
-
-        Returns
-        -------
-        list
-            List of the direct progeny of the nuclide, ordered by decreasing branching
-            fraction.
-
-        Examples
-        --------
-        >>> K40 = rd.Radionuclide('K-40')
-        >>> K40.progeny()
-        ['Ca-40', 'Ar-40']
-
-        """
-
-        return list(self.prog_bf_mode.keys())
-
[…]
