### 2021-09-03 1e65a6c
add nuclide id parser to utils, Inventory id input
