### 2021-09-04 5032e87
add nuclide and util tests for new functionality
