### 2021-09-06 058534a
For release 0.4.0
Tidying up:
- docs for new nuclide module, add utils to API, typos
- changelog
- setup.py license classifier
- ran black
- linted with pylint
- added support for canonical ids in Inventory.remove() method
- tests coverage
