+++ radioactivedecay/inventory.py
@@ -422,9 +422,9 @@ class Inventory:
         Parameters
         ----------
         add_contents : dict
-            Dictionary containing nuclide strings/ids or Nuclide objects as keys
-            and the amount of each nuclide (with specified units) as values
-            which are added to the Inventory.
+            Dictionary containing nuclide strings, canonical ids or Nuclide objects as keys and the
+            amount of each nuclide (with specified units) as values which are added to the
+            Inventory.
         units : str, optional
             Units of the values in the dictionary (e.g. 'Bq', 'Ci', 'g', 'mol', 'num'; 'Bq' is the
             default).
@@ -534,16 +534,16 @@ class Inventory:
 
     @_method_dispatch
     def remove(
-        self, delete: Union[str, Nuclide, List[Union[str, Nuclide]]]
+        self, delete: Union[str, int, Nuclide, List[Union[str, int, Nuclide]]]
     ) -> None:
         """
         Removes nuclide(s) from the inventory.
 
         Parameters
         ----------
-        delete : str or Nuclide or list
-            Nuclide string, Nuclide object or list of nuclide strings or
-            Nuclide objects to delete from the Inventory object.
+        delete : str or int or Nuclide or list
+            Nuclide string, canonical id, Nuclide object or list of nuclide strings, canonical ids
+            or Nuclide objects to delete from the Inventory object.
 
         Examples
         --------
@@ -557,13 +557,29 @@ class Inventory:
 
         """
         raise NotImplementedError(
-            "Method takes a nuclide string, a Nuclide instance, or list of nuclide strings "
-            + "and Nuclide instances as a parameter."
+            "Method takes a nuclide string, a canonical id, a Nuclide instance, or list of nuclide"
+            + " strings, canonical ids or Nuclide instances as a parameter."
         )
 
     @remove.register(str)
     def _(
         self, delete: str
+    ) -> Callable[[Dict[str, float], str, bool, DecayData], None]:
+        """Remove nuclide string from this inventory."""
+
+        delete = parse_nuclide(
+            delete, self.decay_data.nuclides, self.decay_data.dataset_name
+        )
+        new_contents = self.contents.copy()
+        if delete not in new_contents:
+            raise ValueError(delete + " does not exist in this inventory.")
+        new_contents.pop(delete)
+
+        self.contents = new_contents
+
+    @remove.register(int)
+    def _(
+        self, delete: int
     ) -> Callable[[Dict[str, float], str, bool, DecayData], None]:
         """Remove nuclide string from this inventory."""
 
@@ -595,7 +611,7 @@ class Inventory:
 
     @remove.register(list)
     def _(
-        self, delete: List[Union[str, Nuclide]]
+        self, delete: List[Union[str, int, Nuclide]]
     ) -> Callable[[Dict[str, float], str, bool, DecayData], None]:
         """Remove list of nuclide(s) from this inventory."""
 
@@ -799,9 +815,7 @@ class Inventory:
 
         """
 
-        return {
-            nuc: Nuclide(nuc, self.decay_data).progeny() for nuc in self.contents
-        }
+        return {nuc: Nuclide(nuc, self.decay_data).progeny() for nuc in self.contents}
 
     def branching_fractions(self) -> Dict[str, List[float]]:
         """
@@ -849,8 +863,7 @@ class Inventory:
         """
 
         return {
-            nuc: Nuclide(nuc, self.decay_data).decay_modes()
-            for nuc in self.contents
+            nuc: Nuclide(nuc, self.decay_data).decay_modes() for nuc in self.contents
         }
 
     def plot(
@@ -1049,8 +1062,9 @@ class Inventory:
 
 class InventoryHP(Inventory):
     """
[…]
