+++ docs/source/radionuclide.rst
@@ -1,6 +1,0 @@
-radionuclide
-============
-
-.. automodule:: radioactivedecay.radionuclide
-   :members:
-   
