+++ docs/source/utils.rst
@@ -1,0 +1,5 @@
+utils
+=====
+
+.. automodule:: radioactivedecay.utils
+   :members:

[…]
