### 2021-09-07 1de4870
Add ICRP-07 and AMDC license files to MANIFEST.in
