+++ docs/source/conf.py
@@ -24,7 +24,7 @@ copyright = "2021, JAEA"
 author = "Alex Malins"
 
 # The full version, including alpha/beta/rc tags
-release = "0.4.0"
+release = "0.4.1"
 
 
 # -- General configuration ---------------------------------------------------
