+++ MANIFEST.in
@@ -1,4 +1,6 @@
 include LICENSE
+include LICENSE.AMDC
+include LICENSE.ICRP-07
 include CHANGELOG.md
 include CODE_OF_CONDUCT.md
 include CONTRIBUTING.md
