### 2021-09-07 alexmalins/radioactivedecay#37 @munkm
I think if you cut a new release and update the paper with Thom as a co-author items 3 and 4 should be resolved by that. I have no issues with us closing this issue if that's the plan to move forward. 🙂
