### 2021-09-07 alexmalins/radioactivedecay#38 @munkm
Awesome! I think this issue can be closed, then!
