### 2021-09-08 alexmalins/radioactivedecay#38 @alexmalins
Great, thanks!
