### 2021-09-08 alexmalins/radioactivedecay#37 @alexmalins
Hi @munkm,

Yep, that's the plan. The v0.4 release added the new mass and mole fractions functionality. We're just in the process of updating the paper and adding Thom as a new co-author. Please bear with us, will update and close off this issue once done.
