### 2021-09-10 ff987f5
New API methods, changed radionuclide to nuclide
