### 2021-09-13 1eb8647
add InventoryHP discussion, update acknowledgements, minor edits, updates and formatting
