### 2021-09-15 74ab6c7
For release 0.4.2: fix InventoryHP bug affecting decay() method, hard-code Avogadro constant and add tests, update notebooks comparing with PyNE
