### 2021-09-19 #3318 @alexmalins
Ok, everything is now ready. The paper is updated and consistent with the latest release of radioactivedecay (v0.4.2), as are the comparisons with PyNE and Radiological Toolbox. Thom Lemoine is added as a co-author to the paper. Everything that was outstanding in @munkm's review comments ([issue 37](https://github.com/alexmalins/radioactivedecay/issues/37) is closed off.

Do you know what happens with the version number of record (currently v0.3.2, written in first post in this thread by @whedon)? Is it possible to update that to v0.4.2?
