### 2021-09-19 alexmalins/radioactivedecay#37 @alexmalins
Ok, all is ready now. PR #42 updated the paper to the joss_paper branch. Thanks for your helpful suggestions, @munkm !
