### 2021-09-20 #3318 @kellyrowland
Once the review is complete, I'll have the JOSS bot update the version number so that it's up-to-date in the paper. Doing that is part of the check-out process for the editors since it's often the case that software is updated through the paper review.
