### 2021-09-29 83ef13e
Refactor storage of progeny, bfs and decay modes in DecayData, + bug fixes and code quality updates
