+++ tests/test_decaydata.py
@@ -250,20 +250,24 @@ class TestDecayData(unittest.TestCase):
         self.assertEqual(data.nuclides[-1], "He-3")
         self.assertEqual(data.nuclide_dict["Fm-257"], 0)
         self.assertEqual(data.nuclide_dict["He-3"], 1497)
-        self.assertEqual(list(data.prog_bfs_modes[0].keys())[0], "Cf-253")
-        self.assertEqual(list(data.prog_bfs_modes[0].values())[0][0], 0.9979)
-        self.assertEqual(list(data.prog_bfs_modes[0].values())[0][1], "\u03b1")
-        self.assertEqual(list(data.prog_bfs_modes[0].keys())[1], "SF")
-        self.assertEqual(list(data.prog_bfs_modes[0].values())[1][0], 0.0021)
-        self.assertEqual(list(data.prog_bfs_modes[0].values())[1][1], "SF")
-        self.assertEqual(list(data.prog_bfs_modes[-1].keys()), [])
-        self.assertEqual(list(data.prog_bfs_modes[-1].values()), [])
+        self.assertEqual(data.progeny[0][0], "Cf-253")
+        self.assertEqual(data.bfs[0][0], 0.9979)
+        self.assertEqual(data.modes[0][0], "\u03b1")
+        self.assertEqual(data.progeny[0][1], "SF")
+        self.assertEqual(data.bfs[0][1], 0.0021)
+        self.assertEqual(data.modes[0][1], "SF")
+        self.assertEqual(data.progeny[-1], [])
+        self.assertEqual(data.bfs[-1], [])
+        self.assertEqual(data.modes[-1], [])
         self.assertEqual(
             data.scipy_data.decay_consts[0], np.log(2) / (100.5 * 24 * 60 * 60)
         )
         self.assertEqual(data.scipy_data.ln2, np.log(2))
+        self.assertEqual(data.scipy_data.matrix_c.shape, (1498, 1498))
         self.assertEqual(data.scipy_data.matrix_c[0, 0], 1.0)
+        self.assertEqual(data.scipy_data.matrix_c_inv.shape, (1498, 1498))
         self.assertEqual(data.scipy_data.matrix_c_inv[0, 0], 1.0)
+        self.assertEqual(data.scipy_data.matrix_e.shape, (1498, 1498))
         self.assertEqual(data.scipy_data.matrix_e[0, 0], 0.0)
         self.assertEqual(data.scipy_data.vector_n0[0], 0.0)
         self.assertEqual(data.sympy_data, None)
@@ -283,20 +287,24 @@ class TestDecayData(unittest.TestCase):
         self.assertEqual(data.nuclides[-1], "He-3")
         self.assertEqual(data.nuclide_dict["Fm-257"], 0)
         self.assertEqual(data.nuclide_dict["He-3"], 1497)
-        self.assertEqual(list(data.prog_bfs_modes[0].keys())[0], "Cf-253")
-        self.assertEqual(list(data.prog_bfs_modes[0].values())[0][0], 0.9979)
-        self.assertEqual(list(data.prog_bfs_modes[0].values())[0][1], "\u03b1")
-        self.assertEqual(list(data.prog_bfs_modes[0].keys())[1], "SF")
-        self.assertEqual(list(data.prog_bfs_modes[0].values())[1][0], 0.0021)
-        self.assertEqual(list(data.prog_bfs_modes[0].values())[1][1], "SF")
-        self.assertEqual(list(data.prog_bfs_modes[-1].keys()), [])
-        self.assertEqual(list(data.prog_bfs_modes[-1].values()), [])
+        self.assertEqual(data.progeny[0][0], "Cf-253")
[…]
