+++ docs/source/conf.py
@@ -24,7 +24,12 @@ copyright = "2021, JAEA"
 author = "Alex Malins"
 
 # The full version, including alpha/beta/rc tags
-release = "0.4.5"
+with open(f"../../{project}/__init__.py", "r") as file:
+    for line in file:
+        if line.startswith("__version__"):
+            version = line.strip().split()[-1][1:-1]
+            break
+release = version
 
 
 # -- General configuration ---------------------------------------------------
