### 2021-11-07 0779a11
Update README to include correct class
The `Radionuclide` class seems to have been replaced by the `Nuclide` class.
Updated the README to reflect that change in the examples.
