### 2021-11-08 518b692
fix old refs to Radionuclide class, small rewrites in docs
