+++ docs/source/plotting.rst
@@ -4,13 +4,13 @@ Plotting
 Decay chain diagrams
 --------------------
 
-Use the ``Radionuclide`` class ``plot()`` method to create a diagram of the
+Use the ``Nuclide`` class ``plot()`` method to create a diagram of the
 decay chain originating from a radionuclide:
 
 .. code-block:: python3
 
     >>> import radioactivedecay as rd
-    >>> nuc = rd.Radionuclide('Mo-99')
+    >>> nuc = rd.Nuclide('Mo-99')
     >>> fig, ax = nuc.plot()
 
 .. image:: images/Mo-99_chain.png
@@ -27,7 +27,7 @@ default value of 0.5:
 .. code-block:: python3
 
     >>> import radioactivedecay as rd
-    >>> nuc = rd.Radionuclide('Rn-22')
+    >>> nuc = rd.Nuclide('Rn-222')
     >>> fig, ax = nuc.plot(label_pos=0.66)
 
 .. image:: images/Rn-222_chain.png
@@ -37,7 +37,7 @@ Inventory decay graphs
 ----------------------
 
 The ``Inventory`` and ``InventoryHP`` class  ``plot()`` method is for creating
-graphs of the radioactive decay of the radionuclides in an inventory and their
+graphs of the radioactive decay of the nuclides in an inventory and their
 progeny over time. At its simplest, supply the decay timespan to the method:
 
 .. code-block:: python3
@@ -61,8 +61,8 @@ graph to your own needs:
   
 Now we can see the long-lived Pb-210 radionuclide and its progeny, which form
 over a period of months. Large numbers of curves can make the graphs difficult
-to read. Use the ``display`` parameter to specify only the radionuclides you
-want to display. The curves follow the same order as the list you supply:
+to read. Use the ``display`` parameter to specify only the nuclides you wish to
+display. The curves follow the same order as the list you supply:
 
 .. code-block:: python3
 
@@ -71,9 +71,9 @@ want to display. The curves follow the same order as the list you supply:
 .. image:: images/Rn-222_decay_3.png
   :width: 450
 
-By default radionuclides are plotted according to those highest in the decay
-chains downwards. If you wish to display radionuclides in alphabetical order,
-use the ``order`` parameter:
+By default nuclides are plotted according to those highest in the decay chains
+downwards. If you wish to display nuclides in alphabetical order, use the
+``order`` parameter:
 
 .. code-block:: python3
 
