### 2021-11-14 174ef2c
More in docs about Nuclide atomic data methods, contributors update
