### 2021-11-18 f6aef3b
remove unnecessary pass statements
