+++ radioactivedecay/decaydata.py
@@ -40,7 +40,7 @@ from radioactivedecay.utils import parse_nuclide
 try:
     from importlib import resources
 except ImportError:
-    import importlib_resources as resources
+    import importlib_resources as resources  # type:ignore
 
 
 def _csr_matrix_equal(matrix_a: sparse.csr_matrix, matrix_b: sparse.csr_matrix) -> bool:
