### 2021-11-18 80e0680
run black, tweak docstrings
