+++ radioactivedecay/nuclide.py
@@ -1,15 +1,15 @@
 """
-The nuclide module defines the ``Nuclide`` class. Each ``Nuclide``
-instance represents one nuclide, and can be contructed from a nuclide
-string or `zzzaaassss` canonical id. The class properties provide an
-easy way of extracting atomic number and mass information, as well as a
-clean name string. Furthermore, additional methods provide an access
-point for mass data of nuclides, and the decay data of radionuclides,
-if present in a specified dataset. The default decay dataset used if
-none is supplied to the constructor is rd.DEFAULTDATA.
+The nuclide module defines the ``Nuclide`` class. Each ``Nuclide`` instance
+represents one nuclide, and can be contructed from a nuclide string or
+`zzzaaassss` canonical id. The class properties provide an easy way of
+extracting atomic number and mass information, as well as a clean name string.
+Furthermore, additional methods provide an access point for mass data of
+nuclides, and the decay data of radionuclides, if present in a specified
+dataset. The default decay dataset used if none is supplied to the constructor
+is rd.DEFAULTDATA.
 
-The code examples shown in the docstrings assume the
-``radioactivedecay`` package has been imported as:
+The docstring code examples assume that ``radioactivedecay`` has been imported
+as `rd`:
 
 .. highlight:: python
 .. code-block:: python
