### 2021-11-18 90b1762
add changelog entry, add codecov badge token
