### 2021-12-15 9f355c6
new setup config for v0.4.8
