+++ pyproject.toml
@@ -1,0 +1,3 @@
+[build-system]
+requires = ["setuptools>=46.4.0", "wheel"]
+build-backend = 'setuptools.build_meta'
