+++ setup.cfg
@@ -1,0 +1,52 @@
+[metadata]
+name = radioactivedecay
+version = attr: radioactivedecay.__version__
+author = Alex Malins, Thom Lemoine, & contributors
+author_email = github@alexmalinsREMOVETHIS.com
+license = MIT, ICRP-07, AMDC
+license_files = LICENSE, LICENSE.ICRP-07, LICENSE.AMDC
+license_file_content_type = text/markdown; charset=UTF-8
+description = A Python package for radioactive decay modelling that supports 1252 radionuclides, decay chains, branching, and metastable states.
+long_description = file: README.md
+long_description_content_type = text/markdown; charset=UTF-8
+url = https://github.com/radioactivedecay/radioactivedecay
+edit_on_github = True
+github_project = radioactivedecay/radioactivedecay
+project_urls =
+    Bug Tracker = https://github.com/radioactivedecay/radioactivedecay/issues
+    Discussions = https://github.com/radioactivedecay/radioactivedecay/discussions
+    Documentation = https://radioactivedecay.github.io
+    Source Code = https://github.com/radioactivedecay/radioactivedecay
+classifiers = 
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    License :: OSI Approved :: MIT License
+    License :: Other/Proprietary License
+    Operating System :: OS Independent
+    Development Status :: 5 - Production/Stable
+    Intended Audience :: Science/Research
+    Intended Audience :: Education
+    Topic :: Scientific/Engineering
+    Topic :: Education
+
+[options]
+packages = find:
+python_requires = >=3.6
+install_requires =
+    matplotlib
+    networkx
+    numpy
+    scipy
+    sympy
+    importlib_resources; python_version<'3.7'
+
+[options.package_data]
+* = icrp107_ame2020_nubase2020/*.pickle, icrp107_ame2020_nubase2020/*.npz
+
+[coverage:report]
+omit = tests/*
