### 2021-12-15 7b9ea4c
add code coverage GitHub action
