### 2021-12-24 1d7b8db
tweak issue & pr templates
