+++ .github/pull_request_template.md
@@ -1,8 +1,6 @@
-<!--
-Thank you for contributing a Pull Request! Please ensure you check the
-contributor guidelines:
-https://github.com/radioactivedecay/radioactivedecay/blob/main/CONTRIBUTING.md
--->
+<!-- Thank you for contributing a Pull Request! Please ensure you also check the contributor guidelines:
+https://github.com/radioactivedecay/radioactivedecay/blob/main/CONTRIBUTING.md -->
+
 #### What does this PR implement/fix?
 
 
