+++ notebooks/comparisons/README.md
@@ -1,0 +1,4 @@
+# We've moved!
+
+All comparisons notebooks have moved to this
+[repo](https://github.com/radioactivedecay/comparisons).
