### 2022-01-14 #3318 @alexmalins
Hi @kellyrowland @munkm,
This change was just another minor update to URLs in the paper: we moved the [dataset creation](https://github.com/radioactivedecay/datasets) and [code comparison](https://github.com/radioactivedecay/comparisons) notebooks into their own repos.
