+++ radioactivedecay/converters.py
@@ -24,15 +24,13 @@ class UnitConverter(ABC):
     """
     Template class for unit converters using either floats or SymPy arithmetic.
 
-    Parameters
-    ----------
-    year_conv : float or Rational
-        Conversion factor for number of days in a year.
 
-    Attributes
+    Class Attributes
     ----------
     time_units : dict
         Dictionary containing numbers of each time unit per second.
+    year_units : set
+        Set containing all year units.
     activity_units : dict
         Dictionary containing amounts of each activity unit per Bq.
     mass_units : dict
@@ -42,40 +40,31 @@ class UnitConverter(ABC):
 
     """
 
-    def __init__(self, year_conv: Union[float, Rational]):
-        self.time_units = self.get_time_units(year_conv)
-        self.activity_units = self.get_activity_units()
-        self.mass_units = self.get_mass_units()
-        self.moles_units = self.get_moles_units()
-
+    @property
     @abstractmethod
-    def get_time_units(
-        self, year_conv: Union[float, Rational]
-    ) -> Dict[str, Union[float, Expr]]:
-        """
-        Returns time units dictionary (template).
-        """
+    def time_units(self):
+        pass
 
+    year_units = {"y", "yr", "year", "years", "ky", "My", "By", "Gy", "Ty", "Py"}
+
+    @property
     @abstractmethod
-    def get_activity_units(self) -> Dict[str, Union[float, Expr]]:
-        """
-        Returns activity units dictionary (template).
-        """
+    def activity_units(self):
+        pass
 
+    @property
     @abstractmethod
-    def get_mass_units(self) -> Dict[str, Union[float, Expr]]:
-        """
-        Returns mass units dictionary (template).
-        """
+    def mass_units(self):
+        pass
 
+    @property
     @abstractmethod
-    def get_moles_units(self) -> Dict[str, Union[float, Expr]]:
-        """
-        Returns moles units dictionary (template).
-        """
+    def mole_units(self):
+        pass
 
+    @classmethod
     def time_unit_conv(
-        self, time_period: Union[float, Expr], units_from: str, units_to: str
+        cls, time_period: Union[float, Expr], units_from: str, units_to: str, year_conv: Union[float, Expr]
     ) -> Union[float, Expr]:
         """
         Converts a time period from one time unit to another.
@@ -85,13 +74,15 @@ class UnitConverter(ABC):
         time_period : float or Expr
             Time period before conversion.
         units_from : str
-            Time unit before conversion
+            Time unit before conversion.
         units_to : str
-            Time unit after conversion
+            Time unit after conversion.
+        year_conv : Union[float, Expr]
+            Number of days in one year.
 
         Returns
         -------
-        Expr
+        float or Expr
             Time period in new units.
 
         Raises
@@ -101,20 +92,24 @@ class UnitConverter(ABC):
 
         """
 
-        if units_from not in self.time_units:
+        if units_from not in cls.time_units:
             raise ValueError(
                 str(units_from)
                 + ' is not a valid unit, e.g. "s", "m", "h", "d" or "y".'
             )
-        if units_to not in self.time_units:
+        if units_to not in cls.time_units:
             raise ValueError(
[…]
