+++ tests/test_converters.py
@@ -21,7 +21,7 @@ class TestConverters(unittest.TestCase):
 
     def test_avogadro(self) -> None:
         """
-        Test instantiation of UnitConverterFloat objects.
+        Test module constants.
         """
 
         self.assertEqual(AVOGADRO, 6.02214076e23)
@@ -32,18 +32,16 @@ class TestUnitConverterFloat(unittest.TestCase):
     Unit tests for the converters.py UnitConverterFloat class.
     """
 
-    def test_instantiation(self) -> None:
+    def test_class_attributes(self) -> None:
         """
-        Test instantiation of UnitConverterFloat objects.
+        Test class attributes of UnitConverterFloat objects.
         """
 
-        year_conv = 365.2422
-        uconv = UnitConverterFloat(year_conv)
-        self.assertEqual(uconv.time_units["s"], 1.0)
-        self.assertEqual(uconv.time_units["y"], 86400.0 * year_conv)
-        self.assertEqual(uconv.activity_units["Bq"], 1.0)
-        self.assertEqual(uconv.mass_units["g"], 1.0)
-        self.assertEqual(uconv.moles_units["mol"], 1.0)
+        self.assertEqual(UnitConverterFloat.time_units["s"], 1.0)
+        self.assertEqual(UnitConverterFloat.time_units["y"], 86400.0)
+        self.assertEqual(UnitConverterFloat.activity_units["Bq"], 1.0)
+        self.assertEqual(UnitConverterFloat.mass_units["g"], 1.0)
+        self.assertEqual(UnitConverterFloat.moles_units["mol"], 1.0)
 
     def test_time_unit_conv_seconds(self) -> None:
         """
@@ -51,48 +49,47 @@ class TestUnitConverterFloat(unittest.TestCase):
         """
 
         year_conv = 365.2422
-        uconv = UnitConverterFloat(year_conv)
 
-        self.assertEqual(uconv.time_unit_conv(1.0, "s", "s"), 1.0e0)
+        self.assertEqual(UnitConverterFloat.time_unit_conv(1.0, "s", "s", year_conv), 1.0e0)
         self.assertAlmostEqual(
-            uconv.time_unit_conv(1.0, "s", "ps"), 1.0e12, places=(15 - 12)
+            UnitConverterFloat.time_unit_conv(1.0, "s", "ps", year_conv), 1.0e12, places=(15 - 12)
         )
         self.assertAlmostEqual(
-            uconv.time_unit_conv(1.0, "s", "ns"), 1.0e9, places=(15 - 9)
+            UnitConverterFloat.time_unit_conv(1.0, "s", "ns", year_conv), 1.0e9, places=(15 - 9)
         )
-        self.assertEqual(uconv.time_unit_conv(1.0, "s", "μs"), 1.0e6)
-        self.assertEqual(uconv.time_unit_conv(1.0, "s", "us"), 1.0e6)
-        self.assertEqual(uconv.time_unit_conv(1.0, "s", "ms"), 1.0e3)
-        self.assertEqual(uconv.time_unit_conv(1.0, "s", "m"), 1.0 / 60.0)
-        self.assertEqual(uconv.time_unit_conv(1.0, "s", "h"), 1.0 / (60.0 ** 2))
-        self.assertEqual(uconv.time_unit_conv(1.0, "s", "d"), 1.0 / (60.0 ** 2 * 24.0))
[…]
