### 2022-01-30 8f33ba7
fix other code to work with UnitConverter classes
