+++ tests/test_inventory.py
@@ -6,6 +6,10 @@ import copy
 import unittest
 from unittest.mock import patch
 from sympy import Integer, log
+from radioactivedecay.converters import (
+    UnitConverterFloat,
+    UnitConverterSympy
+)
 from radioactivedecay.decaydata import load_dataset, DEFAULTDATA
 from radioactivedecay.inventory import (
     Inventory,
@@ -32,7 +36,6 @@ class TestInventory(unittest.TestCase):
         self.assertEqual(inv.decay_data, DEFAULTDATA)
         self.assertEqual(inv.decay_matrices, DEFAULTDATA.scipy_data)
         self.assertEqual(inv.quantity_converter, DEFAULTDATA.float_quantity_converter)
-        self.assertEqual(inv.unit_converter, DEFAULTDATA.float_unit_converter)
 
         # check instantiation using Nuclide instance
         tritium = Nuclide("H3")
@@ -97,32 +100,32 @@ class TestInventory(unittest.TestCase):
         Test that the contents dictionary values (amounts of each nuclide) are physical.
         """
 
+        inv = Inventory({"H3": 1}, "num", True)
         qconv = DEFAULTDATA.float_quantity_converter
-        uconv = DEFAULTDATA.float_unit_converter
         self.assertEqual(
-            Inventory._convert_to_number({"H-3": 1.0}, "num", qconv, uconv),
+            inv._convert_to_number({"H-3": 1.0}, "num", qconv),
             {"H-3": 1.0},
         )
         self.assertEqual(
-            Inventory._convert_to_number({"H-3": 1.0}, "Bq", qconv, uconv),
+            inv._convert_to_number({"H-3": 1.0}, "Bq", qconv),
             {"H-3": 560892895.7794082},
         )
         self.assertEqual(
-            Inventory._convert_to_number({"H-3": 1.0}, "mBq", qconv, uconv),
+            inv._convert_to_number({"H-3": 1.0}, "mBq", qconv),
             {"H-3": 560892.8957794083},
         )
         self.assertEqual(
-            Inventory._convert_to_number({"H-3": 1.0}, "mol", qconv, uconv),
+            inv._convert_to_number({"H-3": 1.0}, "mol", qconv),
             {"H-3": 6.02214076e23},
         )
         self.assertEqual(
-            Inventory._convert_to_number({"He-3": 1.0}, "kg", qconv, uconv),
+            inv._convert_to_number({"He-3": 1.0}, "kg", qconv),
             {"He-3": 1.9967116089131648e26},
         )
 
         # check catch of incorrect units
         with self.assertRaises(ValueError):
-            Inventory._convert_to_number({"H-3": 1.0}, "xyz", qconv, uconv)
+            inv._convert_to_number({"H-3": 1.0}, "xyz", qconv)
 
     def test_nuclides(self) -> None:
         """
@@ -672,10 +675,9 @@ class TestInventoryHP(unittest.TestCase):
         self.assertEqual(inv.decay_data, DEFAULTDATA)
         self.assertEqual(inv.decay_matrices, DEFAULTDATA.sympy_data)
         self.assertEqual(inv.quantity_converter, DEFAULTDATA.sympy_quantity_converter)
-        self.assertEqual(inv.unit_converter, DEFAULTDATA.sympy_unit_converter)
 
         temp_data = copy.deepcopy(DEFAULTDATA)
-        temp_data.sympy_unit_converter = None
+        temp_data.sympy_year_conv = None
         with self.assertRaises(ValueError):
             InventoryHP({"H3": 1}, "Bq", True, temp_data)
         temp_data.sympy_quantity_converter = None
