+++ radioactivedecay/decaydata.py
@@ -29,6 +29,7 @@ from scipy import sparse
 import sympy
 from sympy import Matrix
 from sympy.matrices import SparseMatrix
+from sympy.core.expr import Expr
 from radioactivedecay.converters import (
     UnitConverterFloat,
     UnitConverterSympy,
@@ -272,10 +273,10 @@ class DecayMatricesSympy(DecayMatrices):
 class DecayData:
     """
     Instances of DecayData store a complete radioactive decay dataset. It stores data for decay
-    calculations, inventory transformations and unit conversions using DecayMatrices,
-    UnitConverterFloat, QuantityConverter, DecayMatricesSympy, UnitConverterSympy and
-    QuantityConverterSympy objects. Other decay data which are reported to users but not used in
-    calculations are stored in other instance attributes.
+    calculations, inventory transformations and quantity conversions using DecayMatrices,
+    QuantityConverter, DecayMatricesSympy, and QuantityConverterSympy objects. Other decay data
+    which are reported to users but not used in calculations are stored in other instance
+    attributes.
 
     Parameters
     ----------
@@ -283,6 +284,8 @@ class DecayData:
         Name of the decay dataset.
     bfs : numpy.ndarray
         NumPy array of lists with branching fraction data.
+    float_year_conv : float
+        Number of days in one year for float time unit conversions.
     hldata : numpy.ndarray
         List of tuples containing half-life floats, time unit strings and readable format half-life
         strings.
@@ -294,14 +297,11 @@ class DecayData:
         NumPy array of lists with direct progeny data.
     scipy_data : DecayMatrices
         Dataset of double precision decay matrices (SciPy/NumPy objects).
-    float_unit_converter : UnitConverterFloat
-        Convert between units of a single quantity using floating point operations.
     sympy_data : None or DecayMatricesSympy
         Dataset of arbitrary-precision decay matrices (SymPy objects). Or None if SymPy
         functionality is not used. Default is None.
-    sympy_unit_converter : None or UnitConverterSympy
-        Convert between units of a single quantity using SymPy arbitrary precision operations.
-        Default is None.
+    sympy_year_conv : None or Expr
+        Number of days in one year for SymPy time unit conversions.
 
     Attributes
     ----------
@@ -311,8 +311,8 @@ class DecayData:
         Name of the decay dataset.
     float_quantity_converter : QuantityConverter
         Convert between quantities using floating point operations.
-    float_unit_converter : UnitConverterFloat
-        Convert between units of a single quantity using floating point operations.
+    float_year_conv : float
+        Number of days in one year for float time unit conversions.
     hldata : numpy.ndarray
         List of tuples containing half-life floats, time unit strings and readable format half-life
         strings.
@@ -331,8 +331,8 @@ class DecayData:
         is not used.
     sympy_quantity_converter : None or QuantityConverterSympy
         Convert between quantities using SymPy arbitrary precision operations.
-    sympy_unit_converter : None or UnitConverterSympy
-        Convert between units of a single quantity using SymPy arbitrary precision operations.
+    sympy_year_conv : None or Expr
+        Number of days in one year for SymPy time unit conversions.
 
     """
 
@@ -340,19 +340,19 @@ class DecayData:
         self,
         dataset_name: str,
         bfs: np.ndarray,
+        float_year_conv: float,
         hldata: np.ndarray,
         modes: np.ndarray,
         nuclides: np.ndarray,
         progeny: np.ndarray,
         scipy_data: DecayMatrices,
[…]
