### 2022-01-30 880f5cb
Refactor QuantityConverter classes
