+++ radioactivedecay/converters.py
@@ -102,10 +102,13 @@ class UnitConverter(ABC):
                 str(units_to) + ' is not a valid unit, e.g. "s", "m", "h", "d" or "y".'
             )
 
-        result = time_period * cls.time_units[units_from] / cls.time_units[units_to]
-        if units_from in cls.year_units: result *= year_conv
-        if units_to in cls.year_units: result /= year_conv
-        return result
+        factor_from = cls.time_units[units_from]
+        factor_to = cls.time_units[units_to]
+        if units_from in cls.year_units:
+            factor_from *= year_conv
+        if units_to in cls.year_units:
+            factor_to *= year_conv
+        return time_period * factor_from / factor_to
 
     @classmethod
     def activity_unit_conv(
@@ -315,6 +318,10 @@ class UnitConverterFloat(UnitConverter):
         "Mmol": 1.0e6,
     }
 
+    @classmethod
+    def __repr__(cls) -> str:
+        return f"{cls.__name__} using double-precision floats."
+
 
 class UnitConverterSympy(UnitConverter):
     """
@@ -403,58 +410,41 @@ class UnitConverterSympy(UnitConverter):
         "Mmol": Integer(1000000),
     }
 
+    @classmethod
+    def __repr__(cls) -> str:
+        return f"{cls.__name__} using SymPy arbitrary precision calculations."
 
 
-class QuantityConverter:
+class QuantityConverter(ABC):
     """
-    Converts activity in Bq, mass in g and moles in mol to number of atoms, and vice versa.
-    Supports both Exprs or SymPy quantities.
+    Template class for quantity converters using either floats or SymPy arithmetic. Converts
+    activity in Bq, mass in g and moles in mol to number of atoms, and vice versa.
 
-    Parameters
+    Class Attributes
     ----------
-    atomic_masses : numpy.ndarray
-        Column vector of the atomic masses (in g/mol).
-    decay_consts : numpy.ndarray
-        Column vector of the decay constants (in s\\ :sup:`-1`).
-    nuclide_dict : dict
-        Dictionary containing nuclide strings as keys and positions in the matrices as values.
-
-    Attributes
-    ----------
-    atomic_masses : numpy.ndarray
-        Column vector of the atomic masses (in g/mol).
-    avogadro : float
+    avogadro : float or Expr
         Avogadro constant (number of atoms/mol).
-    decay_consts : numpy.ndarray
-        Column vector of the decay constants (in s\\ :sup:`-1`).
-    nuclide_dict : dict
-        Dictionary containing nuclide strings as keys and positions in the matrices as values.
 
     """
 
-    def __init__(
-        self,
-        nuclide_dict: Dict[str, int],
-        atomic_masses: np.ndarray,
-        decay_consts: np.ndarray,
-    ) -> None:
-        self.nuclide_dict = nuclide_dict
-        self.atomic_masses = atomic_masses
-        self.decay_consts = decay_consts
-        self.avogadro = AVOGADRO
+    @property
+    @abstractmethod
+    def time_units(self):
+        pass
 
+    @staticmethod
     def activity_to_number(
-        self, nuclide: str, activity: Union[float, Expr]
+       activity: Union[float, Expr], decay_const: Union[float, Expr]
     ) -> Union[float, Expr]:
         """
         Converts an activity in Bq to the number of atoms.
 
         Parameters
         ----------
-        nuclide : str
-            Nuclide string.
         activity : float or sympy.core.expr.Expr
             The activity in Bq of the nuclide to be converted.
+        decay_const : float or sympy.core.expr.Expr
+            Decay constant of the nuclide in s^-1.
 
         Returns
         -------
@@ -463,20 +453,21 @@ class QuantityConverter:
 
         """
 
[…]
