+++ tests/test_converters.py
@@ -9,7 +9,7 @@ from radioactivedecay.converters import (
     AVOGADRO,
     UnitConverterFloat,
     UnitConverterSympy,
-    QuantityConverter,
+    QuantityConverterFloat,
     QuantityConverterSympy,
 )
 
@@ -309,6 +309,15 @@ class TestUnitConverterFloat(unittest.TestCase):
         with self.assertRaises(ValueError):
             UnitConverterFloat.moles_unit_conv(1.0, "tmol", 1.0)
 
+    def test___repr__(self) -> None:
+        """
+        Test UnitConverterFloat __repr__ strings.
+        """
+
+        self.assertEqual(
+            UnitConverterFloat.__repr__(), "UnitConverterFloat using double-precision floats."
+        )
+
 
 class TestUnitConverterSympy(unittest.TestCase):
     """
@@ -516,143 +525,83 @@ class TestUnitConverterSympy(unittest.TestCase):
         with self.assertRaises(ValueError):
             UnitConverterSympy.moles_unit_conv(Integer(1), "tmol", Integer(1))
 
-
-class TestQuantityConverter(unittest.TestCase):
-    """
-    Unit tests for the converters.py QuantityConverter class.
-    """
-
-    def test_instantiation(self) -> None:
+    def test___repr__(self) -> None:
         """
-        Test instantiation of QuantityConverter objects.
+        Test UnitConverterSympy __repr__ strings.
         """
 
-        nuclide_dict = {"H-3": 0, "He-3": 1}
-        atomic_masses = np.array([3.01604928132, 3.01602932197])
-        decay_consts = np.array([1.7828715741004621e-09, 0.0])
-        qconv = QuantityConverter(nuclide_dict, atomic_masses, decay_consts)
-        self.assertEqual(qconv.nuclide_dict["He-3"], 1)
-        self.assertEqual(qconv.atomic_masses[0], 3.01604928132)
-        self.assertEqual(qconv.decay_consts[1], 0)
-        self.assertEqual(qconv.avogadro, 6.02214076e23)
+        self.assertEqual(
+            UnitConverterSympy.__repr__(), "UnitConverterSympy using SymPy arbitrary precision calculations."
+        )
+
+
+class TestQuantityConverterFloat(unittest.TestCase):
+    """
+    Unit tests for the converters.py QuantityConverterFloat class.
+    """
+
+    def test_class_attribute(self) -> None:
+        """
+        Test class attribute of QuantityConverterFloat.
+        """
+
+        self.assertEqual(QuantityConverterFloat.avogadro, 6.02214076e23)
 
     def test_activity_to_number(self) -> None:
         """
         Test the conversion of activity in Bq to number of atoms.
         """
 
-        nuclide_dict = {"H-3": 0, "He-3": 1}
-        atomic_masses = np.array([3.01604928132, 3.01602932197])
-        decay_consts = np.array([1.7828715741004621e-09, 0.0])
-        qconv = QuantityConverter(nuclide_dict, atomic_masses, decay_consts)
-        self.assertEqual(qconv.activity_to_number("H-3", 1.0), 560892895.7794082)
+        self.assertEqual(QuantityConverterFloat.activity_to_number(1.0, 1.7828715741004621e-09), 560892895.7794082)
 
     def test_mass_to_number(self) -> None:
         """
         Test the conversion of mass in grams to number of atoms.
         """
 
-        nuclide_dict = {"H-3": 0, "He-3": 1}
-        atomic_masses = np.array([3.01604928132, 3.01602932197])
-        decay_consts = np.array([1.7828715741004621e-09, 0.0])
-        qconv = QuantityConverter(nuclide_dict, atomic_masses, decay_consts)
-
-        self.assertEqual(qconv.mass_to_number("H-3", 1.0), 1.996698395247825e23)
[…]
