### 2022-01-30 f0ebb72
fix rest to work with new QuantityConverter class
