### 2022-01-30 8695a8c
Improve docstrings/code coverage
