+++ tests/test_converters.py
@@ -3,7 +3,6 @@ Unit tests for converts.py classes and methods.
 """
 
 import unittest
-import numpy as np
 from sympy import Integer, log
 from radioactivedecay.converters import (
     AVOGADRO,
@@ -50,37 +49,75 @@ class TestUnitConverterFloat(unittest.TestCase):
 
         year_conv = 365.2422
 
-        self.assertEqual(UnitConverterFloat.time_unit_conv(1.0, "s", "s", year_conv), 1.0e0)
-        self.assertAlmostEqual(
-            UnitConverterFloat.time_unit_conv(1.0, "s", "ps", year_conv), 1.0e12, places=(15 - 12)
+        self.assertEqual(
+            UnitConverterFloat.time_unit_conv(1.0, "s", "s", year_conv), 1.0e0
         )
         self.assertAlmostEqual(
-            UnitConverterFloat.time_unit_conv(1.0, "s", "ns", year_conv), 1.0e9, places=(15 - 9)
+            UnitConverterFloat.time_unit_conv(1.0, "s", "ps", year_conv),
+            1.0e12,
+            places=(15 - 12),
         )
-        self.assertEqual(UnitConverterFloat.time_unit_conv(1.0, "s", "μs", year_conv), 1.0e6)
-        self.assertEqual(UnitConverterFloat.time_unit_conv(1.0, "s", "us", year_conv), 1.0e6)
-        self.assertEqual(UnitConverterFloat.time_unit_conv(1.0, "s", "ms", year_conv), 1.0e3)
-        self.assertEqual(UnitConverterFloat.time_unit_conv(1.0, "s", "m", year_conv), 1.0 / 60.0)
-        self.assertEqual(UnitConverterFloat.time_unit_conv(1.0, "s", "h", year_conv), 1.0 / (60.0 ** 2))
-        self.assertEqual(UnitConverterFloat.time_unit_conv(1.0, "s", "d", year_conv), 1.0 / (60.0 ** 2 * 24.0))
+        self.assertAlmostEqual(
+            UnitConverterFloat.time_unit_conv(1.0, "s", "ns", year_conv),
+            1.0e9,
+            places=(15 - 9),
+        )
+        self.assertEqual(
+            UnitConverterFloat.time_unit_conv(1.0, "s", "μs", year_conv), 1.0e6
+        )
+        self.assertEqual(
+            UnitConverterFloat.time_unit_conv(1.0, "s", "us", year_conv), 1.0e6
+        )
+        self.assertEqual(
+            UnitConverterFloat.time_unit_conv(1.0, "s", "ms", year_conv), 1.0e3
+        )
+        self.assertEqual(
+            UnitConverterFloat.time_unit_conv(1.0, "s", "m", year_conv), 1.0 / 60.0
+        )
+        self.assertEqual(
+            UnitConverterFloat.time_unit_conv(1.0, "s", "h", year_conv),
+            1.0 / (60.0 ** 2),
+        )
+        self.assertEqual(
+            UnitConverterFloat.time_unit_conv(1.0, "s", "d", year_conv),
+            1.0 / (60.0 ** 2 * 24.0),
+        )
         self.assertEqual(
             UnitConverterFloat.time_unit_conv(1.0, "s", "y", year_conv),
             1.0 / (60.0 ** 2 * 24.0 * year_conv),
         )
         self.assertAlmostEqual(
-            UnitConverterFloat.time_unit_conv(1.0, "ps", "s", year_conv), 1.0e-12, places=(12 + 15)
+            UnitConverterFloat.time_unit_conv(1.0, "ps", "s", year_conv),
+            1.0e-12,
+            places=(12 + 15),
[…]
