### 2022-01-31 af3518f
change property abstractmethods to class attributes to fix mypy errors
