### 2022-01-31 79a68c8
fix some inventory mypy errors
