### 2022-01-31 f019b74
change some attributes to properties so as not to store same data twice
