+++ radioactivedecay/nuclide.py
@@ -49,23 +49,8 @@ class Nuclide:
     ----------
     nuclide: str
         Nuclide name string.
-    Z : int
-        Atomic number.
-    A : int
-        Atomic mass number.
-    id : int
-        Canonical nuclide id, in zzzaaassss form. Ground state is 0000,
-        first excited state ("m") is 0001, second ("n") is 0002, etc.
     decay_data : DecayData
         Decay dataset.
-    atomic_mass : float
-        Atomic weight of the nuclide, in g/mol.
-    prog : list
-        List of direct progeny of this nuclide.
-    bfs : list
-        List of branching fractions to direct progeny of this nuclide.
-    modes : list
-        List of modes to direct progeny of this nuclide.
 
     Examples
     --------
@@ -84,60 +69,53 @@ class Nuclide:
         self, input_nuclide: Union[str, int], decay_data: DecayData = DEFAULTDATA
     ) -> None:
         self.decay_data = decay_data
-        self.parse_name(input_nuclide)
-        idx = decay_data.nuclide_dict[self.nuclide]
-        self.atomic_mass = decay_data.scipy_data.atomic_masses[idx]
-        self.prog: List[str] = decay_data.progeny[idx]
-        self.bfs: List[float] = decay_data.bfs[idx]
-        self.modes: List[str] = decay_data.modes[idx]
-
-    def parse_name(self, input_nuclide: Union[str, int]) -> None:
-        """
-        Parse input and set atomic data attributes.
-        """
-
         self.nuclide = parse_nuclide(
             input_nuclide, self.decay_data.nuclides, self.decay_data.dataset_name
         )
-        self.Z = elem_to_Z(self.nuclide.split("-")[0])
-        self.A = int(self.nuclide.split("-")[1].strip("mn"))
-        self.state = self.nuclide.split("-")[1].strip("0123456789")
-        self.id = build_id(self.Z, self.A, self.state)
 
-    def __repr__(self) -> str:
-        rep = (
-            "Nuclide: "
-            + self.nuclide
-            + ", decay dataset: "
-            + self.decay_data.dataset_name
-        )
-
-        return rep
-
-    def __eq__(self, other: object) -> bool:
+    @property
+    def Z(self) -> int:
         """
-        Check whether two ``Nuclide`` instances are equal with ``==`` operator.
+        Returns the atomic number of the nuclide.
         """
 
-        if not isinstance(other, Nuclide):
-            return NotImplemented
-        return self.nuclide == other.nuclide and self.decay_data == other.decay_data
+        return elem_to_Z(self.nuclide.split("-")[0])
 
-    def __ne__(self, other: object) -> bool:
+    @property
+    def A(self) -> int:
         """
-        Check whether two ``Nuclide`` instances are not equal with ``!=`` operator.
-        """
-
-        if not isinstance(other, Nuclide):
-            return NotImplemented
-        return not self.__eq__(other)
-
-    def __hash__(self) -> int:
-        """
-        Hash function for ``Nuclide`` instances.
+        Returns the atomic mass number of the nuclide.
         """
 
-        return hash((self.nuclide, self.decay_data.dataset_name))
+        return int(self.nuclide.split("-")[1].strip("mn"))
+
+    @property
+    def state(self) -> str:
+        """
+        Returns the excited state letter. '' for ground state, 'm' for first excited state, 'n' for
+        second excited state, etc.
+        """
+
+        return self.nuclide.split("-")[1].strip("0123456789")
+
+    @property
+    def id(self) -> int:
+        """
+        Returns the canonical nuclide id, in zzzaaassss form. Ground state is 0000, first excited
+        state ("m") is 0001, second ("n") is 0002, etc.
+        """
[…]
