### 2022-01-31 415925b
improve Nuclide property docstrings + other docs improvements
