### 2022-01-31 568fb97
Adds another Nuclide.half_life() docstring example
