### 2022-01-31 d7cc310
add changelog entry, improve DecayData docstring
