### 2022-02-06 f721fed
fix module docstring formatting, suppress some mypy errors
