### 2022-02-06 25afc88
Make better Table of Contents for API section of docs
