+++ docs/source/decaydata.rst
@@ -2,4 +2,25 @@ decaydata
 =========
 
 .. automodule:: radioactivedecay.decaydata
+
+DecayData
+---------
+
+.. autoclass:: radioactivedecay.decaydata.DecayData
    :members:
+
+DecayMatrices
+-------------
+
+.. autoclass:: radioactivedecay.decaydata.DecayMatrices
+   :show-inheritance:
+   :members:
+   :inherited-members:
+
+DecayMatricesSympy
+------------------
+
+.. autoclass:: radioactivedecay.decaydata.DecayMatricesSympy
+   :show-inheritance:
+   :members:
+   :inherited-members:
