### 2022-02-06 79621c1
Large refactor for inventory module - Clearer module docstring - Use inbuilt singledispatchmethod for Python >3.7 - Make AbstractInventory ABC to base Inventory and InventoryHP off (template design pattern) - Fix a number of mypy errors and suppress others - plots module decay_graph method is made public - Suppress line-too-long pylint warnings
