+++ radioactivedecay/inventory.py
@@ -1,13 +1,13 @@
 """
-The inventory module defines the ``Inventory`` class. An ``Inventory`` instance contains one or
-more nuclides each with an associated amount (number of atoms). The decay of the nuclide(s) in an
-``Inventory`` can be calculated by using the ``decay()`` method (normal double-precision
-floating-point operations). The corresponding ``InventoryHP`` class performs decay calculations
-with SymPy high numerical precision operations. A ``DecayData`` dataset is associated with
-``Inventory`` and ``InventoryHP`` instances (default is rd.DEFAULTDATA).
+The inventory module defines the ``Inventory`` and ``InventoryHP`` classes. Instances of these
+classes contain one or more nuclides each with an associated amount (number of atoms). The decay
+of the nuclide(s) can be calculated by using the ``decay()`` method. The ``Inventory`` class
+performs calculations using double-precision floats. The ``InventoryHP`` class performs
+calculations using SymPy high numerical precision operations. A ``DecayData`` dataset is associated
+with ``Inventory`` and ``InventoryHP`` instances (default is rd.DEFAULTDATA).
 
 The docstring code examples assume that ``radioactivedecay`` has been imported
-as `rd`:
+as ``rd``:
 
 .. highlight:: python
 .. code-block:: python
@@ -16,7 +16,7 @@ as `rd`:
 
 """
 
-from functools import singledispatch, update_wrapper
+from abc import ABC, abstractmethod
 from typing import Any, Callable, Dict, List, Optional, Tuple, Type, Union
 import matplotlib
 import numpy as np
@@ -24,8 +24,10 @@ from scipy import sparse
 from sympy import exp, Integer, Matrix, nsimplify
 from sympy.core.expr import Expr
 from radioactivedecay.converters import (
+    QuantityConverter,
     QuantityConverterFloat,
     QuantityConverterSympy,
+    UnitConverter,
     UnitConverterFloat,
     UnitConverterSympy,
 )
@@ -35,7 +37,7 @@ from radioactivedecay.decaydata import (
     DecayMatricesSympy,
     DEFAULTDATA,
 )
-from radioactivedecay.plots import _decay_graph
+from radioactivedecay.plots import decay_graph
 from radioactivedecay.nuclide import Nuclide
 from radioactivedecay.utils import (
     parse_nuclide,
@@ -44,27 +46,35 @@ from radioactivedecay.utils import (
     sort_list_according_to_dataset,
 )
 
+# functools.singledispatchmethod is new in Python version 3.8
+# this block adds support manually for Python <3.8
+# the except block can be removed once min required Python version is 3.8
+try:
+    from functools import singledispatchmethod
+except ImportError:
+    from functools import singledispatch, update_wrapper
+
+    def singledispatchmethod(func: Callable[..., Any]):  # type:ignore
+        """Adds singledispatch support for instance methods of a class."""
+
+        dispatcher = singledispatch(func)
+
+        def wrapper(*args, **kwargs):  # type:ignore
+            return dispatcher.dispatch(args[1].__class__)(*args, **kwargs)
+
+        wrapper.register = dispatcher.register
+        update_wrapper(wrapper, func)
+        return wrapper
+
 
 # pylint: disable=too-many-arguments, too-many-lines, too-many-locals
 
 
-def _method_dispatch(func: Callable[..., Any]):
-    """Adds singledispatch support for class methods."""
-
-    dispatcher = singledispatch(func)
-
-    def wrapper(*args, **kwargs):
-        return dispatcher.dispatch(args[1].__class__)(*args, **kwargs)
-
-    wrapper.register = dispatcher.register
-    update_wrapper(wrapper, func)
-    return wrapper
-
-
-class Inventory:
+class AbstractInventory(ABC):
     """
[…]
