### 2022-02-06 ef496a9
Add CHANGELOG entry, clarify decay graph in README
