+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
 ## [0.X.X] - 20XX-XX-XX
-- Code refactoring: reduce coupling between modules by refactoring Converter & Nuclide classes to
-not store duplicate data, but to receive needed data only via their API calls. DecayData objects no
-longer need to store Converter objects. Continuing to reduce pylint/mypy errors/warnings, however
-will need to wait to bump requirement to Python 3.7+ to improve type hinting of NumPy/SciPy arrays
-(requires NumPy 1.21+, holding off for now to maintain Python 3.6 support) (#64).
+- Code refactoring: reduce coupling between modules by refactoring Converter, Nuclide & Inventory
+classes to not store duplicate data, but to receive data when needed via their API calls. DecayData objects
+no longer store Converter objects. Continuing to reduce pylint/mypy errors/warnings, however will
+need to wait to bump requirement to Python 3.7+ to improve type hinting of NumPy/SciPy arrays
+(requires NumPy 1.21+, holding off for now to maintain Python 3.6 support) (#64, #66).
+- Improve API documentation table of contents & README fix (#66)
 - Move dataset creation and comparison notebooks to separate repos within the radioactivedecay org
 on GitHub (#63).
 - Tweak GitHub Issue & PR markdown templates (#62).
