### 2022-02-07 d6b9594
Use template pattern for DecayMatrices - clarify requirement for imporlib_resouces dependency - better error handling if a DecayData instance does not contain SymPy data - more comprehensive decaydata module tests - some docstring fixes
