+++ radioactivedecay/decaydata.py
@@ -1,8 +1,8 @@
 """
-The decaydata module defines the ``DecayData``, ``DecayMatrices`` and ``DecayMatricesSympy``
-classes. ``DecayMatrices`` and ``DecayMatricesSympy`` instances store data used for decay
+The decaydata module defines the ``DecayData``, ``DecayMatricesScipy`` and ``DecayMatricesSympy``
+classes. ``DecayMatricesScipy`` and ``DecayMatricesSympy`` instances store data used for decay
 calculations and other inventory transformations (using double-precision floating point and SymPy
-arbitrary precision numbers, respectively). ``DecayData`` instances store ``DecayMatrices``,
+arbitrary precision numbers, respectively). ``DecayData`` instances store ``DecayMatricesScipy``,
 ``DecayMatricesSympy`` and unit converter objects, along with other decay data for reporting to
 users (this data is not used for calculations).
 
@@ -21,6 +21,7 @@ DEFAULTDATA : DecayData
     decay data and AME2020 and Nubase2020 for atomic mass data.
 """
 
+from abc import ABC, abstractmethod
 from pathlib import Path
 import pickle
 from typing import Any, ContextManager, Optional, Union
@@ -33,6 +34,10 @@ from sympy.core.expr import Expr
 from radioactivedecay.converters import UnitConverterFloat
 from radioactivedecay.utils import parse_nuclide
 
+
+# importlib.resources is new in Python version 3.7
+# this block adds support using a backport for Python <3.7
+# the except block and dependency can be removed once min required Python version is 3.7
 try:
     from importlib import resources
 except ImportError:
@@ -41,8 +46,8 @@ except ImportError:
 
 def _csr_matrix_equal(matrix_a: sparse.csr_matrix, matrix_b: sparse.csr_matrix) -> bool:
     """
-    Checks whether two SciPy Compressed Sparse Row (CSR) matrices are equal (i.e. they have the
-    same elements and all elements are equal).
+    Utility function checking whether two SciPy Compressed Sparse Row (CSR) matrices are equal
+    (i.e. they have the same elements and all elements are equal).
 
     Parameters
     ----------
@@ -64,10 +69,101 @@ def _csr_matrix_equal(matrix_a: sparse.csr_matrix, matrix_b: sparse.csr_matrix) 
     )
 
 
-class DecayMatrices:
+class DecayMatrices(ABC):
     """
-    Instances of DecayMatrices store vectors with fundamental decay/atomic data and matrices used
-    for decay calculations.
+    Template class for for SciPy (float) or SymPy DecayMatrices classes. Instances of these classes
+    store vectors and matrices of data used for decay calculations.
+
+    Parameters
+    ----------
+    atomic_masses : numpy.ndarray or sympy.matrices.dense.MutableSparseMatrix
+        Column vector of the atomic masses (in g/mol).
+    decay_consts : numpy.ndarray or sympy.matrices.dense.MutableSparseMatrix
+        Column vector of the decay constants (in s\\ :sup:`-1`).
+    matrix_c : scipy.sparse.csr.csr_matrix or sympy.matrices.dense.MutableSparseMatrix
+        A pre-calculated sparse lower traingular matrix used in decay calculations.
+    matrix_c_inv : scipy.sparse.csr.csr_matrix or sympy.matrices.dense.MutableSparseMatrix
+        The inverse of matrix_c, also used in decay calculations.
+
+    Attributes
+    ----------
+    atomic_masses : numpy.ndarray or sympy.matrices.dense.MutableSparseMatrix
+        Column vector of the atomic masses (in g/mol).
+    decay_consts : numpy.ndarray or sympy.matrices.dense.MutableSparseMatrix
+        Column vector of the decay constants (in s\\ :sup:`-1`).
[…]
