### 2022-03-04 #3318 @kellyrowland
@alexmalins this all looks great. 🙂 I would have loved to have a tool like this to check my own decay calculations.
