### 2022-03-09 #3318 @kellyrowland
Hi @alexmalins there's a bit of inconsistency with the editoralbot behavior that the JOSS team is working to resolve; the `recommend-acceptance` command is currently failing because of some warnings. The intended behavior is that the `generate pdf` command should also be failing, so we don't get to this mismatched point.

The repeated warnings in the acceptance pipeline look to be all of the form:
```Warning:  Could not convert TeX math \mathbf{N}(t) = C e^{\varLambda_{d} t} C^{-1} \mathbf{N}(0).
  \label{eq:solution}, rendering as TeX:
  ) = C e^{\varLambda_{d} t} C^{-1} \mathb
                     ^
  unexpected control sequence \varLambda
  ```

I'm not seeing `varLambda` listed in http://tug.ctan.org/info/symbols/comprehensive/symbols-a4.pdf ; maybe `upLambda` (see Table 191) would be a sufficient replacement?
