### 2022-03-13 #3318 @xuanxu
@alexmalins That `\varLambda` error is a problem in the [TeXMath package](https://github.com/jgm/texmath) used to parse math. The underlying issue is it's no longer enough for math to show up correctly in PDF, but that it must also be parseable by pandoc.

There's now a PR to fix `\varLambda` support in TeXMath (https://github.com/jgm/texmath/pull/190). It's been merged, so I guess the problem will be solved in the next release.
For the time being, the following could be used as a workaround (best placed close to the top of the Markdown file):
`\def\varLambda{\mathit{\Lambda}}`
