### 2022-03-14 0a9b6f9
Attempt to fix \varLambda issue, update acknowledgements
