### 2022-03-15 309201c
Fix incorrect parsing of SymPy version string This caused import failure with SymPy >=1.10
