+++ docs/source/installation.rst
@@ -7,7 +7,8 @@ Prerequisites
 ``radioactivedecay`` requires Python 3.6+ and the `Matplotlib
 <https://matplotlib.org/>`_, `NetworkX
 <https://networkx.org/>`_, `NumPy <https://numpy.org/>`_,
-`SciPy <https://www.scipy.org/index.html>`_  and 
+`SciPy <https://www.scipy.org/index.html>`_,
+`Setuptools <https://setuptools.pypa.io/en/latest/>`_ and 
 `SymPy <https://www.sympy.org>`_ packages. These can be
 installed from `python.org <https://www.python.org/>`_ and `PyPI
 <https://pypi.org/>`_, or via a package manager such as `Anaconda
@@ -32,7 +33,8 @@ or via `conda-forge <https://anaconda.org/conda-forge/radioactivedecay>`_:
     $ conda install -c conda-forge radioactivedecay
 
 Either command will attempt to install the dependencies (Matplotlib, NetworkX,
-NumPy, SciPy & SymPy) if they are not already present in the environment.
+NumPy, SciPy, Setuptools & SymPy) if they are not already present in the
+environment.
 
 It is also possible to clone the GitHub `repository 
 <https://github.com/radioactivedecay/radioactivedecay>`_ and install from within the
