### 2022-03-15 #3318 @kthyng
@alexmalins I'm taking over to finish the publishing process. Everything looks good except — in the references there are asterisks at the end of the title for several, for example Huang et al. Is there a reason for them to be there?

And hey I went to Whitman too @lemointm!
