### 2022-03-16 1557f38
remove asterisks from three reference titles
