+++ paper/paper.bib
@@ -63,7 +63,7 @@ url = {https://www.ans.org/pubs/transactions/article-38022/}
 
 @article{Huang2021,
 author = {W.J. Huang and Meng Wang and F.G. Kondev and G. Audi and S. Naimi},
-title = {The {AME} 2020 atomic mass evaluation (I). Evaluation of input data, and adjustment procedures{\ast}},
+title = {The {AME} 2020 atomic mass evaluation (I). Evaluation of input data, and adjustment procedures},
 journal = {Chinese Physics C},
 volume = {45},
 number = {3},
@@ -96,7 +96,7 @@ doi = {10.1016/j.icrp.2008.10.003}
 
 @article{Kondev2021,
 author = {F.G. Kondev and M. Wang and W.J. Huang and S. Naimi and G. Audi},
-title = {The {NUBASE}2020 evaluation of nuclear physics properties {\ast}},
+title = {The {NUBASE}2020 evaluation of nuclear physics properties},
 journal = {Chinese Physics C},
 volume = {45},
 number = {3},
@@ -149,7 +149,7 @@ doi = {10.1038/s41592-019-0686-2}
 
 @article{Wang2021,
 author = {Meng Wang and W.J. Huang and F.G. Kondev and G. Audi and S. Naimi},
-title = {The {AME} 2020 atomic mass evaluation ({II}). Tables, graphs and references{\ast}},
+title = {The {AME} 2020 atomic mass evaluation ({II}). Tables, graphs and references},
 journal = {Chinese Physics C},
 volume = {45},
 number = {3},
