### 2022-03-16 #3318 @alexmalins
Hi @kthyng - thanks for checking this paper! Those asterisks came in the .bib files from the journal's website. They needn't be there, so this last commit removed them.
